Crate libtegra

Modules

actmon
Driver for the Tegra X1 Activity Monitor (ACTMON).

ahb
Driver for the Tegra X1 AMBA Advanced High-performance Bus.

apb
Driver for the Tegra X1 AMBA Peripheral Bus Controller.

arm
Utilities related to the ARM processors found on Tegra X1 SoC.

atomic
Abstractions for the Atomics interface provided by the Tegra X1.

bpmp
Abstractions over the Boot and Power Management Processor.

car
Abstractions over the Tegra X1 Clock and Reset Controller functionality.

dsi
Driver for the MIPI Display Serial Interface (DSI) implemented by the Tegra X1.

flow
Driver for the Flow Controller of the Tegra X1.

fuse
Driver for the Tegra X1 FUSE Controller.

gpio
Driver for the Tegra X1 General-Purpose Input/Output controller.

hal
Miscellaneous embedded-hal trait implementations.

i2c
Driver for the Tegra X1 Inter-Integrated Circuit Controller and its Devices.

kfuse
Driver for the Tegra X1 for the HDCP KFUSE Controller.

mc
Driver for the Tegra X1 Memory Controller.

memory_map
Unified Tegra X1 Memory Address Map constants.

pinmux
Driver for Tegra X1 Multi-Purpose Pins and Pin Multiplexing.

pmc
Driver for the Tegra X1 Power Management Controller.

pwm
Driver for the Pulse Width Modulator Controller of the Tegra X1.

se
Driver for the Tegra X1 Security Engine.

spi
Driver for the Tegra X1 Serial Peripheral Interface Controller.

system
System-related APIs for identifying Tegra hardware.

timer
Abstractions of the timer interfaces provided by the Tegra X1.

tsec
Driver for interfacing with the Tegra Security Co-Processor (TSEC).

uart
Driver for the Tegra X1 Universal Asynchronous Receiver/Transmitter Controller.

vic
Driver for interfacing with the Tegra X1 Video Image Compositor (VIC).

Macros

tegra_gpio
A macro to facilitate the creation of a GPIO given port and pin.