Module libtegra::actmon

Driver for the Tegra X1 Activity Monitor (ACTMON).

See Chapter 10 in the Tegra X1 Technical Reference Manual for details.

Description

The Activity Monitor acts as a repository for activity indicators from various units, performs averaging of past samplse, and indicates any abnormality in activity levels by sending an interrupt to software. Devices provide an idle/busy signal to the ACTMON block.

Modules

ACTMON_APB_CTRL_SAPB_0
Bitfields of the ACTMON_APB_CTRL_SAPB_0 register.

ACTMON_CTRL_0
Bitfields of the ACTMON_CPU_CTRL_0 register.

ACTMON_GLB_PERIOD_CTRL_0
Bitfields of the ACTMON_GLB_PERIOD_CTRL_0 register.

ACTMON_GLB_STATUS_0
Bitfields of the ACTMON_GLB_STATUS_0 register.

ACTMON_HISTOGRAM_CONFIG_0
Bitfields of the ACTMON_HISTOGRAM_CONFIG_0 register.

ACTMON_INTR_STATUS_0
Bitfields of the ACTMON_CPU_INTR_STATUS_0 register.

Structs

Registers
Representation of the Activity Montior registers.

Constants

REGISTERS
A pointer to the Activity Monitor registers that can be accessed by dereferencing it.