Module libtegra::actmon::ACTMON_CTRL_0::AHB_MONITOR_COND

Selection criteria on parent module for the activity signal.

Only ever set this in AHB control registers!

Enums

Value
Selection criteria on parent module for the activity signal.

Constants

AhbDma

All

ApbDma

Arc

BSea

CLEAR

CSite

Cop

Cpu

Dds

Na1

Na2

Na3

Na4

Na5

Na6

Na7

Na8

Na9

SET

Se

Usb

Usb2