Module libtegra::actmon::ACTMON_CTRL_0::APB_MONITOR_COND

Selection criteria on parent module for the activity signal.

Only ever set this in APB control registers!

Enums

Value

Selection criteria on parent module for the activity signal.

Constants

All

CLEAR

Dtv

Dvc

I2c1

I2c2

I2c3

I2c4

I2c6

QSpi

Reserved1

Reserved2

SET

Spi1

Spi2

Spi3

Spi4

Spi5

Spi6

UartA

UartB

UartC

UartD