Module libtegra::apb::dma

Driver for the Tegra X1 APB DMA Controller.

See Chapter 21.2 in the Tegra X1 Technical Reference Manual for details.

Modules

APBDMACHAN_CHANNEL_AHB_PTR_0
Bitfields of the APBDMACHAN_CHANNEL_AHB_PTR_0 register.

APBDMACHAN_CHANNEL_AHB_SEQ_0
Bitfields of the APBDMACHAN_CHANNEL_AHB_SEQ_0 register.

APBDMACHAN_CHANNEL_APB_PTR_0
Bitfields of the APBDMACHAN_CHANNEL_APB_PTR_0 register.

APBDMACHAN_CHANNEL_APB_SEQ_0
Bitfields of the APBDMACHAN_CHANNEL_APB_SEQ_0 register.

APBDMACHAN_CHANNEL_CSRE_0
Bitfields of the APBDMACHAN_CHANNEL_CSRE_0 register.

APBDMACHAN_CHANNEL_CSR_0
Bitfields of the APBDMACHAN_CHANNEL_CSR_0 register.

APBDMACHAN_CHANNEL_DMA_BYTE_STA_0
Bitfields of the APBDMACHAN_CHANNEL_DMA_BYTE_STA_0 register.

APBDMACHAN_CHANNEL_STA_0
Bitfields of the APBDMACHAN_CHANNEL_STA_0 register.

APBDMACHAN_CHANNEL_WCOUNT_0
Bitfields of the APBDMACHAN_CHANNEL_WCOUNT_0 register.

APBDMA_CHANNEL_EN_REG_0
Bitfields of the APBDMA_CHANNEL_EN_REG_0 register.

APBDMA_CHANNEL_SWID1_0
Bitfields of the APBDMA_CHANNEL_SWID1_0 register.

APBDMA_CHANNEL_SWID_0
Bitfields of the APBDMA_CHANNEL_SWID_0 register.

APBDMA_CHANNEL_TRIG_REG_0
Bitfields of the APBDMA_CHANNEL_TRIG_REG_0 register.

APBDMA_CHAN_WT_REG0_0
Bitfields of the APBDMA_CHAN_WT_REG0_0 register.

APBDMA_CHAN_WT_REG1_0
Bitfields of the APBDMA_CHAN_WT_REG1_0 register.

APBDMA_CHAN_WT_REG2_0
Bitfields of the APBDMA_CHAN_WT_REG2_0 register.

APBDMA_CHAN_WT_REG3_0
Bitfields of the APBDMA_CHAN_WT_REG3_0 register.

APBDMA_CNTRL_REG_0
Bitfields of the APBDMA_CNTRL_REG_0 register.

APBDMA_COMMAND_0
Bitfields of the APBDMA_COMMAND_0 register.

APBDMA_DMA_STATUS_0
Bitfields of the APBDMA_DMA_STATUS_0 register.

APBDMA_IRQ_MASK_0
Bitfields of the APBDMA_IRQ_MASK_0 register.

APBDMA_IRQ_MASK_CLR_0
Bitfields of the APBDMA_IRQ_MASK_CLR_0 register.

APBDMA_IRQ_MASK_SET_0
Bitfields of the APBDMA_IRQ_MASK_SET_0 register.

APBDMA_IRQ_STA_COP_0
Bitfields of the APBDMA_IRQ_STA_COP_0 register.

APBDMA_IRQ_STA_CPU_0
Bitfields of the APBDMA_IRQ_STA_CPU_0 register.

APBDMA_SECURITY_REG_0
Bitfields of the APBDMA_SECURITY_REG_0 register.

APBDMA_STATUS_0
Bitfields of the APBDMA_STATUS_0 register.

APBDMA_TRIG_REG_0
Bitfields of the APBDMA_TRIG_REG_0 register.

DMACHAN_CHANNEL_WORD_TRANSFER_0
Bitfields of the DMACHAN_CHANNEL_WORD_TRANSFER_0 register.

Structs

Channel
Representation of an APB DMA Channel.

ChannelRegisters
Representation of the APB DMA Channel registers.

Controller
Representation of the AMBA Peripheral Bus DMA Controller.

Registers
Representation of the APB DMA registers.

Constants

CHANNEL_0
A pointer to the APB DMA Channel 0 register block that can be accessed by dereferencing it.

CHANNEL_1
A pointer to the APB DMA Channel 1 register block that can be accessed by dereferencing it.

CHANNEL_2
A pointer to the APB DMA Channel 2 register block that can be accessed by dereferencing it.

CHANNEL_3
A pointer to the APB DMA Channel 3 register block that can be accessed by dereferencing it.

CHANNEL_4
A pointer to the APB DMA Channel 4 register block that can be accessed by dereferencing it.

CHANNEL_5
A pointer to the APB DMA Channel 5 register block that can be accessed by dereferencing it.

CHANNEL_6
A pointer to the APB DMA Channel 6 register block that can be accessed by dereferencing it.

CHANNEL_7
A pointer to the APB DMA Channel 7 register block that can be accessed by dereferencing it.

CHANNEL_8
A pointer to the APB DMA Channel 8 register block that can be accessed by dereferencing it.

CHANNEL_9
A pointer to the APB DMA Channel 9 register block that can be accessed by dereferencing it.

CHANNEL_10
A pointer to the APB DMA Channel 10 register block that can be accessed by dereferencing it.

CHANNEL_11
A pointer to the APB DMA Channel 11 register block that can be accessed by dereferencing it.

CHANNEL_12
A pointer to the APB DMA Channel 12 register block that can be accessed by dereferencing it.

CHANNEL_13
A pointer to the APB DMA Channel 13 register block that can be accessed by dereferencing it.

CHANNEL_14
A pointer to the APB DMA Channel 14 register block that can be accessed by dereferencing it.

CHANNEL_15
A pointer to the APB DMA Channel 15 register block that can be accessed by dereferencing it.

CHANNEL_16
A pointer to the APB DMA Channel 16 register block that can be accessed by dereferencing it.

CHANNEL_17
A pointer to the APB DMA Channel 17 register block that can be accessed by dereferencing it.

CHANNEL_18
A pointer to the APB DMA Channel 18 register block that can be accessed by dereferencing it.

CHANNEL_19
A pointer to the APB DMA Channel 19 register block that can be accessed by dereferencing it.

CHANNEL_20
A pointer to the APB DMA Channel 20 register block that can be accessed by dereferencing it.

CHANNEL_21
A pointer to the APB DMA Channel 21 register block that can be accessed by dereferencing it.

CHANNEL_22
A pointer to the APB DMA Channel 22 register block that can be accessed by dereferencing it.

CHANNEL_23
A pointer to the APB DMA Channel 23 register block that can be accessed by dereferencing it.

CHANNEL_24
A pointer to the APB DMA Channel 24 register block that can be accessed by dereferencing it.

CHANNEL_25
A pointer to the APB DMA Channel 25 register block that can be accessed by dereferencing it.

CHANNEL_26
A pointer to the APB DMA Channel 26 register block that can be accessed by dereferencing it.

CHANNEL_27
A pointer to the APB DMA Channel 27 register block that can be accessed by dereferencing it.

CHANNEL_28
A pointer to the APB DMA Channel 28 register block that can be accessed by dereferencing it.

CHANNEL_29
A pointer to the APB DMA Channel 29 register block that can be accessed by dereferencing it.

CHANNEL_30
A pointer to the APB DMA Channel 30 register block that can be accessed by dereferencing it.

CHANNEL_31
A pointer to the APB DMA Channel 31 register block that can be accessed by dereferencing it.

REGISTERS
A pointer to the APB DMA register block that can be accessed by dereferencing it.