Struct libtegra::apb::dma::Controller

pub struct Controller { /* fields omitted */ }

Representation of the AMBA Peripheral Bus DMA Controller.

The controller manages 32 DMA Channels, which are used to transfer data over DMA. Various bus protocols support a DMA mode for data transfers, which can be implemented by interfacing with this device.

Implementations

impl Controller

pub const fn new() -> Self

Creates a new instance of the APB DMA Controller.

NOTE: Please refrain from calling this method multiple times. It is advised to create a single, global instance of the controller and stick to using it.

pub unsafe fn get_channel(&mut self, index: usize) -> &mut Channel

Unconditionally returns a reference to the channel with the supplied index.

Panics

If the given index is not within the range of available channels (0-31), the method will panic.

Safety

This method is considered unsafe because it is within the caller's responsibility to ensure that the channel is correctly acquired and released and that no running DMA transfers on the channel are being corrupted through interaction with it.

pub fn find_free_channel(&mut self) -> Option<&mut Channel>

Finds a currently unoccupied channel and returns it.

May be None if all channels are busy.

pub fn execute_in_channel<C>(
&mut self,
channel: Option<Channel>,
f: C
) -> Result<(), ()> where
C: FnOnce(&Channel) -> Result<(), ()>,

Reserves a Channel for use and passes its reference through the supplied closure.

This method gives users the possibility to correctly acquire and release channels respectively, without having to worry about unwanted side effects, such as race conditions. If a specific channel is desired, it can be passed directly to this method, otherwise an unclaimed channel will be randomly picked.

The supplied closure takes the Channel as argument and is expected to return a Result, which is forwarded to the direct return value of this method. Note that Channel::start and Channel::finish must be called by the user within the closure to allow more freedom in how to configure and use the DMA interfaces of specific devices.

Example

use libtegra::apb;

/// The global instance of the APB DMA Controller.
const APB_DMA_CONTROLLER: apb::dma::Controller = apb::dma::Controller::new();

// Do something with DMA Channel 0...
APB_DMA_CONTROLLER.execute_in_channel(Some(apb::dma::Channel::CH0), |channel| {
    // Within this context, we exclusively own Channel 0, which can
    // be accessed through `channel` of type `&apb::dma::Channel`.

    // Do something with channel...

    Ok(())
})
.unwrap();

// DMA Channel 0 was released, it can be re-used now.