Module libtegra::arm::gic

Implementation of the ARM Generic Interrupt Controller v2.

Modules

gicc
Representation of the CPU Interface registers for the GIC.

gicd
Representation of the Distributor registers for the GIC.

Structs

Gic
Representation of the Generic Interrupt Controller.

IrqContext
Interrupt context token.

IrqDescriptor
Describes an IRQ with a corresponding handler function.

IrqNumber
A wrapper for IRQ numbers that enforces the number to be in a valid range.

Enums

IrqMode
Enum for selecting in which mode to trigger interrupts on.

Type Definitions

Irq
An interrupt request for the GIC.

IrqHandler
A generic callback that acts as an interrupt handler.