Module libtegra::arm::gic::gicc

Representation of the CPU Interface registers for the GIC.

Access to these registers is banked so that each CPU core sees their own instance of the GICC when accessing the same registers. It is used to acknowledge and handle pending IRQs.

Modules

GICC_AHPPIR

Bitfields of the GICC_AHPPIR register.

GICC_AIAR

Bitfields of the GICC_AIAR register.

GICC_BPR

Bitfields of the GICC_BPR register.

GICC_CTLR

Bitfields of the GICC_CTLR register.

GICC_EOIR

Bitfields of the GICC_EOIR register.

GICC_HPPIR

Bitfields of the GICC_HPPIR register.

GICC_IAR

Bitfields of the GICC_IAR register.

GICC_IIDR

Bitfields of the GICC_IIDR register.

GICC_PMR

Bitfields of the GICC_PMR register.

Structs

Registers

Representation of the GIC CPU Interface registers.