Struct libtegra::arm::gic::Gic

pub struct Gic { /* fields omitted */ }

Representation of the Generic Interrupt Controller.

Implementations

impl Gic

pub const unsafe fn get(gicc_base: usize, gicd_base: usize) -> Self

Gets an instance of the GIC through the base addresses to the CPU Interface and Distributor registers.

Safety

The caller is responsible for providing correct physical addresses pointing to the GICC and GICD devices. Otherwise, MMIO access hits uncontrolled addresses and undefined behavior will be the result.

pub fn init(&self)

Initializes and enables the GIC device for interrupt delivery.

NOTE: This method must be called once before the GIC is usable.

pub fn get_num_irqs(&mut self) -> usize

Gets the number of IRQs implemented in hardware.

pub fn get_pending_irq<'ctx>(&'ctx self, _ic: &IrqContext<'ctx>) -> Irq

Extracts the IRQ with the highest priority from pending interrupts.

This function should be directly called from the CPU's IRQ exception vector. A reference to IrqContext must be passed to this method to ensure that this is the case.

pub fn mark_irq_completed<'ctx>(&'ctx self, irq: Irq, _ic: &IrqContext<'ctx>)

Completes the handling of a currently active IRQ by signaling an EOI (End of Interrupt).

This function should be directly called from the CPU's IRQ exception vector. A reference to IrqContext must be passed to this method to ensure that this is the case.

pub fn register_descriptor(&mut self, irq: Irq, descriptor: IrqDescriptor)

Registers an interrupt descriptor that corresponds to the given IRQ number.

This overrides previously configured IRQ descriptors for the interrupt, if any.

pub fn dispatch_irq<'ctx>(
&'ctx self,
irq: Irq,
_ic: &IrqContext<'ctx>
) -> Result<(), &'static str>

Dispatches the interrupt handler corresponding to the given IRQ.

This function should be directly called from the CPU's IRQ exception vector. A reference to IrqContext must be passed to this method to ensure that this is the case.

Any potential errors when no interrupt handler is actually registered or when the handler itself encountered an error will be consumed into the Result returned by this function.

pub fn set_irq_enable(&self, irq: Irq, enable: bool)

Sets the enabled state for a specific IRQ.

By marking an interrupt as disabled, it will not be dispatched by the GIC anymore until it is re-enabled again and vice versa. Can be used to bypass specific IRQs either permanently or temporarily.

pub fn set_spi_mode(&self, irq: Irq, mode: IrqMode)

Configures the interrupt triggering mode for the given SPI.

pub fn set_irq_pending(&self, irq: Irq)

Sets a given IRQ to a pending state in the GIC.

pub fn set_irq_priority(&self, irq: Irq, priority: u32)

Sets the delivery priority for a given IRQ.

The lower the value, the higher the priority. 0 is the highest interrupt delivery priority supported by the GIC.

pub fn set_irq_group(&self, irq: Irq, group: u32)

Assigns a new group to a given IRQ.

On boot or reset, all SPIs are in group 0, making them secure interrupt. Putting individual SPIs into group 1 effectively makes them secure interrupts.

pub fn set_spi_target_cpu(&self, irq: Irq, cpu: u32)

Configures the target CPU core to route the given IRQs to.