Module libtegra::dsi

Driver for the MIPI Display Serial Interface (DSI) implemented by the Tegra X1.

Modules

DSI_CTXSW_0
Bitfields of the DSI_CTXSW_0 register.

DSI_DSI_BTA_TIMING_0
Bitfields of the DSI_DSI_BTA_TIMING_0 register.

DSI_DSI_CONTROL_0
Bitfields of the DSI_DSI_CONTROL_0 register.

DSI_DSI_DCS_CMDS_0
Bitfields of the DSI_DSI_DCS_CMDS_0 register.

DSI_DSI_DSC_CONTROL_0
Bitfields of the DSI_DSI_DSC_CONTROL_0 register.

DSI_DSI_GANGED_MODE_CONTROL_0
Bitfields of the DSI_DSI_GANGED_MODE_CONTROL_0 register.

DSI_DSI_GANGED_MODE_SIZE_0
Bitfields of the DSI_DSI_GANGED_MODE_SIZE_0 register.

DSI_DSI_GANGED_MODE_START_0
Bitfields of the DSI_DSI_GANGED_MODE_START_0 register.

DSI_DSI_INIT_SEQ_CONTROL_0
Bitfields of the DSI_DSI_INIT_SEQ_CONTROL_0 register.

DSI_DSI_MAX_THRESHOLD_0
Bitfields of the DSI_DSI_MAX_THRESHOLD_0 register.

DSI_DSI_PHY_TIMING_0_0
Bitfields of the DSI_DSI_PHY_TIMING_0_0 register.

DSI_DSI_PHY_TIMING_1_0
Bitfields of the DSI_DSI_PHY_TIMING_1_0 register.

DSI_DSI_PHY_TIMING_2_0
Bitfields of the DSI_DSI_PHY_TIMING_2_0 register.

DSI_DSI_PKT_LEN_0_1_0
Bitfields of the DSI_DSI_PKT_LEN_0_1_0 register.

DSI_DSI_PKT_LEN_2_3_0
Bitfields of the DSI_DSI_PKT_LEN_2_3_0 register.

DSI_DSI_PKT_LEN_4_5_0
Bitfields of the DSI_DSI_PKT_LEN_4_5_0 register.

DSI_DSI_PKT_LEN_6_7_0
Bitfields of the DSI_DSI_PKT_LEN_6_7_0 register.

DSI_DSI_PKT_SEQ_0_HI_0
Bitfields of the DSI_DSI_PKT_SEQ_0_HI_0 register.

DSI_DSI_PKT_SEQ_0_LO_0
Bitfields of the DSI_DSI_PKT_SEQ_0_LO_0 register.

DSI_DSI_PKT_SEQ_1_HI_0
Bitfields of the DSI_DSI_PKT_SEQ_1_HI_0 register.

DSI_DSI_PKT_SEQ_1_LO_0
Bitfields of the DSI_DSI_PKT_SEQ_1_LO_0 register.

DSI_DSI_PKT_SEQ_2_HI_0
Bitfields of the DSI_DSI_PKT_SEQ_2_HI_0 register.

DSI_DSI_PKT_SEQ_2_LO_0
Bitfields of the DSI_DSI_PKT_SEQ_2_LO_0 register.

DSI_DSI_PKT_SEQ_3_HI_0
Bitfields of the DSI_DSI_PKT_SEQ_3_HI_0 register.

DSI_DSI_PKT_SEQ_3_LO_0
Bitfields of the DSI_DSI_PKT_SEQ_3_LO_0 register.

DSI_DSI_PKT_SEQ_4_HI_0
Bitfields of the DSI_DSI_PKT_SEQ_4_HI_0 register.

DSI_DSI_PKT_SEQ_4_LO_0
Bitfields of the DSI_DSI_PKT_SEQ_4_LO_0 register.

DSI_DSI_PKT_SEQ_5_HI_0
Bitfields of the DSI_DSI_PKT_SEQ_5_HI_0 register.

DSI_DSI_PKT_SEQ_5_LO_0
Bitfields of the DSI_DSI_PKT_SEQ_5_LO_0 register.

DSI_DSI_POWER_CONTROL_0
Bitfields of the DSI_DSI_POWER_CONTROL_0 register.

DSI_DSI_RAW_DATA_BYTE_COUNT_0
Bitfields of the DSI_DSI_RAW_DATA_BYTE_COUNT_0 register.

DSI_DSI_SOL_DELAY_0
Bitfields of the DSI_DSI_SOL_DELAY_0 register.

DSI_DSI_STATUS_0
Bitfields of the DSI_DSI_STATUS_0 register.

DSI_DSI_TIMEOUT_0_0
Bitfields of the DSI_DSI_TIMEOUT_0_0 register.

DSI_DSI_TIMEOUT_1_0
Bitfields of the DSI_DSI_TIMEOUT_1_0 register.

DSI_DSI_TO_TALLY_0
Bitfields of the DSI_DSI_TO_TALLY_0 register.

DSI_DSI_TRIGGER_0
Bitfields of the DSI_DSI_TRIGGER_0 register.

DSI_DSI_ULTRA_LOW_POWER_CONTROL_0
Bitfields of the DSI_DSI_ULTRA_LOW_POWER_CONTROL_0 register.

DSI_DSI_VID_MODE_CONTROL_0
Bitfields of the DSI_DSI_VID_MODE_CONTROL_0 register.

DSI_DUMMY_PIX_CNT_0
Bitfields of the DSI_DUMMY_PIX_CNT_0 register.

DSI_HOST_DSI_CONTROL_0
Bitfields of the DSI_HOST_DSI_CONTROL_0 register.

DSI_INCR_SYNCPT_0
Bitfields of the DSI_INCR_SYNCPT_0 register.

DSI_INCR_SYNCPT_CNTRL_0
Bitfields of the DSI_INCR_SYNCPT_CNTRL_0 register.

DSI_INCR_SYNCPT_ERROR_0
Bitfields of the DSI_INCR_SYNCPT_ERROR_0 register.

DSI_INT_ENABLE_0
Bitfields of the DSI_INT_ENABLE_0 register.

DSI_INT_MASK_0
Bitfields of the DSI_INT_MASK_0 register.

DSI_INT_STATUS_0
Bitfields of the DSI_INT_STATUS_0 register.

DSI_PAD_CD_STATUS_0
Bitfields of the DSI_PAD_CD_STATUS_0 register.

DSI_PAD_CONTROL_0
Bitfields of the DSI_PAD_CONTROL_0 register.

DSI_PAD_CONTROL_1_0
Bitfields of the DSI_PAD_CONTROL_1_0 register.

DSI_PAD_CONTROL_2_0
Bitfields of the DSI_PAD_CONTROL_2_0 register.

DSI_PAD_CONTROL_3_0
Bitfields of the DSI_PAD_CONTROL_3_0 register.

DSI_PAD_CONTROL_4_0
Bitfields of the DSI_PAD_CONTROL_4_0 register.

DSI_PAD_CONTROL_CD_0
Bitfields of the DSI_PAD_CONTROL_CD_0 register.

Structs

Registers
Representation of the MIPI-DSI registers.

Constants

DSIA_REGISTERS
A pointer to the DSIA register block that can be accessed by dereferencing it.

DSIB_REGISTERS
A pointer to the DSIB register block that can be accessed by dereferencing it.