Module libtegra::gpio

Driver for the Tegra X1 General-Purpose Input/Output controller.

See Chapter 9.10 in the Tegra X1 Technical Reference Manual for details.

Description

The GPIO controller for Tegra X1 devices provides the tools for configuring each MPIO for use as a software-controlled GPIO.

Controller Structure

The controller is divided into 8 banks. Each bank contains 4 GPIO ports, which in turn provide 8 available pins, numbered from 0 through 7. GPIO ports a labeled consecutively from A through Z and then AA through FF. Ports A through E are in bank 0, E through H in bank 1, and so on.

In conclusion, 256 GPIOs in total are therefore available, approximately 170 of them being physical pins. Each of them can be identified uniquely through a port and a pin.

Configuration

Each GPIO can be configured individually as Input/Output/Interrupt sources with level/edge controls.

use libtegra::{tegra_gpio, gpio};

// This...
tegra_gpio!(D, 4).config(gpio::Config::OutputHigh);

// ...is equivalent to:
let pin = gpio::Gpio {
    port: gpio::Port::D,
    pin: gpio::Pin::P4,
};
pin.set_mode(gpio::Mode::Gpio);
pin.set_direction(gpio::Direction::Output);
pin.write(gpio::Level::High);

Structs

Gpio
Representation of a Tegra X1 GPIO.

GpioBank
Representation of a GPIO bank.

GpioController
Representation of the GPIO controller.

Enums

Config
Supported GPIO configurations.

Direction
Possible GPIO directions.

InterruptType
Supported interrupt types.

Level
Possible GPIO levels.

Mode
Possible GPIO modes.

Pin
The GPIO pins that are provided for each port.

Port
The GPIO ports that are supported by the platform.

Constants

CONTROLLER
A pointer to the GPIO controller that can be accessed by dereferencing it.