Module libtegra::i2c

Driver for the Tegra X1 Inter-Integrated Circuit Controller and its Devices.

See Chapter 35 in the Tegra X1 Technical Reference Manual for details.

Description

The I²C controller (I2C) provides a specification-compliant I²C 3.0 driver that supports multiple controllers and devices.

There are six instances of the I2C controller within Tegra X1 devices, all of them providing the same functionality. These six instances are represented as publicly exposed constants of the I2c structure.

Transmission speed

By default, the I2C controller instances which are exposed by this module have pre-defined clocks, which are configured for 100KHz data transfer rate over the I²C protocol.

If customized clocks are needed, one must create a custom instance of an I2C controller with the desired Clock settings.

Implementation details

The implementation of the I2c controller provides lower-level and higher-level methods for driving the host MMIO driver in Normal Mode with 7-bit addressing transactions. That being said, if one wishes to use Packet Mode (which may be added in the future) or 10-bit addressing transactions, one would have to implement the functionality themselves.

Modules

I2C_FIFO_CONTROL_0
Bitfields of the I2C_FIFO_CONTROL_0 register.

I2C_FIFO_STATUS_0
Bitfields of the I2C_FIFO_STATUS_0 register.

I2C_I2C_BUS_CLEAR_CONFIG_0
Bitfields of the I2C_I2C_BUS_CLEAR_CONFIG_0 register.

I2C_I2C_BUS_CLEAR_STATUS_0
Bitfields of the I2C_I2C_BUS_CLEAR_STATUS_0 register.

I2C_I2C_CLK_DIVISOR_REGISTER_0
Bitfields of the I2C_I2C_CLK_DIVISOR_REGISTER_0 register.

I2C_I2C_CMD_ADDR0_0
Bitfields of the I2C_I2C_CMD_ADDR0_0 register.

I2C_I2C_CMD_ADDR1_0
Bitfields of the I2C_I2C_CMD_ADDR1_0 register.

I2C_I2C_CMD_DATA1_0
Bitfields of the I2C_I2C_CMD_DATA1_0 register.

I2C_I2C_CMD_DATA2_0
Bitfields of the I2C_I2C_CMD_DATA2_0 register.

I2C_I2C_CNFG_0
Bitfields of the I2C_I2C_CNFG_0 register.

I2C_I2C_CONFIG_LOAD_0
Bitfields of the I2C_I2C_CONFIG_LOAD_0 register.

I2C_I2C_HS_INTERFACE_TIMING_0_0
Bitfields of the I2C_I2C_HS_INTERFACE_TIMING_0_0 register.

I2C_I2C_HS_INTERFACE_TIMING_1_0
Bitfields of the I2C_I2C_HS_INTERFACE_TIMING_1_0 register.

I2C_I2C_INTERFACE_TIMING_0_0
Bitfields of the I2C_I2C_INTERFACE_TIMING_0_0 register.

I2C_I2C_INTERFACE_TIMING_1_0
Bitfields of the I2C_I2C_INTERFACE_TIMING_1_0 register.

I2C_I2C_INTERRUPT_SET_REGISTER_0
Bitfields of the I2C_I2C_INTERRUPT_SET_REGISTER_0 register.

I2C_I2C_INTERRUPT_SOURCE_REGISTER_0
Bitfields of the I2C_I2C_INTERRUPT_SOURCE_REGISTER_0 register.

I2C_I2C_RX_FIFO_0
Bitfields of the I2C_I2C_RX_FIFO_0 register.

I2C_I2C_SLV_PACKET_STATUS_0
Bitfields of the I2C_I2C_SLV_PACKET_STATUS_0 register.

I2C_I2C_SLV_RX_FIFO_0
Bitfields of the I2C_I2C_SLV_RX_FIFO_0 register.

I2C_I2C_SLV_TX_PACKET_FIFO_0
Bitfields of the I2C_I2C_SLV_TX_PACKET_FIFO_0 register.

I2C_I2C_SL_ADDR1_0
Bitfields of the I2C_I2C_SL_ADDR1_0 register.

I2C_I2C_SL_ADDR2_0
Bitfields of the I2C_I2C_SL_ADDR2_0 register.

I2C_I2C_SL_CNFG_0
Bitfields of the I2C_I2C_SL_CNFG_0 register.

I2C_I2C_SL_DELAY_COUNT_0
Bitfields of the I2C_I2C_SL_DELAY_COUNT_0 register.

I2C_I2C_SL_INT_MASK_0
Bitfields of the I2C_I2C_SL_INT_MASK_0 register.

I2C_I2C_SL_INT_SET_0
Bitfields of the I2C_I2C_SL_INT_SET_0 register.

I2C_I2C_SL_INT_SOURCE_0
Bitfields of the I2C_I2C_SL_INT_SOURCE_0 register.

I2C_I2C_SL_RCVD_0
Bitfields of the I2C_I2C_SL_RCVD_0 register.

I2C_I2C_SL_STATUS_0
Bitfields of the I2C_I2C_SL_STATUS_0 register.

I2C_I2C_STATUS_0
Bitfields of the I2C_I2C_STATUS_0 register.

I2C_I2C_TLOW_SEXT_0
Bitfields of the I2C_I2C_TLOW_SEXT_0 register.

I2C_I2C_TX_PACKET_FIFO_0
Bitfields of the I2C_I2C_TX_PACKET_FIFO_0 register.

I2C_INTERRUPT_MASK_REGISTER_0
Bitfields of the I2C_INTERRUPT_MASK_REGISTER_0 register.

I2C_INTERRUPT_STATUS_REGISTER_0
Bitfields of the I2C_INTERRUPT_STATUS_REGISTER_0 register.

I2C_PACKET_TRANSFER_STATUS_0
Bitfields of the I2C_PACKET_TRANSFER_STATUS_0 register.

Structs

I2c
Representation of an I2C controller device.

Registers
Representation of the I2C registers.

Enums

Error
I2C errors that may occur during communication over the I²C protocol.

Constants

I2C_1_REGISTERS
A pointer to the I2C 1 register block that can be accessed by dereferencing it.

I2C_2_REGISTERS
A pointer to the I2C 2 register block that can be accessed by dereferencing it.

I2C_3_REGISTERS
A pointer to the I2C 3 register block that can be accessed by dereferencing it.

I2C_4_REGISTERS
A pointer to the I2C 4 register block that can be accessed by dereferencing it.

I2C_5_REGISTERS
A pointer to the I2C 5 register block that can be accessed by dereferencing it.

I2C_6_REGISTERS
A pointer to the I2C 6 register block that can be accessed by dereferencing it.