Module libtegra::i2c::I2C_I2C_BUS_CLEAR_CONFIG_0::BC_TERMINATE

Enums

Value

Constants

CLEAR

Immediate

Terminate the bus clear operation immediately when SDA is released or threshold count is reached, whichever is earlier.

SET

Threshold

Irrespective of SDA release status during bus clear, terminate the bus clear only after the threshold is reached.