Module libtegra::i2c::I2C_I2C_CNFG_0

Bitfields of the I2C_I2C_CNFG_0 register.

Modules

A_MOD
The address mode, defining how the slave address should be programmed.

CMD1
Read/Write Command for Slave 1.

CMD2
Read/Write Command for Slave 2.

DEBOUNCE_CNT
Debounce period for SDA and SCL lines.

LENGTH
The number of bytes to be transmitted per transaction.

MSTR_CLR_BUS_ON_TIMEOUT
If set, the master will put a STOP condition on the bus.

MULTI_MASTER_MODE
Used to select single or multi master mode.

NEW_MASTER_FSM
Whether the new Master FSM should be enabled.

NOACK
Whether an ACK at the end of the disable should be ignored.

PACKET_MODE_EN
Whether transfer should be done in packet mode.

SEND
Setting this bit makes the master initialize transaction in normal mode.

SLV2
Whether two-slave transaction should be enabled.

START
Whether sending a Start byte is required.

Structs

Register

Constants

A_MOD

CMD1

CMD2

DEBOUNCE_CNT

LENGTH

MSTR_CLR_BUS_ON_TIMEOUT

MULTI_MASTER_MODE

NEW_MASTER_FSM

NOACK

PACKET_MODE_EN

SEND

SLV2

START