Module libtegra::i2c::I2C_PACKET_TRANSFER_STATUS_0

Bitfields of the I2C_PACKET_TRANSFER_STATUS_0 register.

Modules

ARB_LOST
Arbitration lost for the current byte.

CONTROLLER_BUSY
Controller is busy.

NOACK_FOR_ADDR
No ACK received for the address byte.

NOACK_FOR_DATA
No ACK received for the data byte.

TRANSFER_BYTENUM
The number of bytes transferred in the current packet.

TRANSFER_COMPLETE
The packet transfer for which the last packet is set has been completed.

TRANSFER_PKT_ID
The current packet ID for which the transaction is happening on the bus.

Structs

Register

Constants

ARB_LOST

CONTROLLER_BUSY

NOACK_FOR_ADDR

NOACK_FOR_DATA

TRANSFER_BYTENUM

TRANSFER_COMPLETE

TRANSFER_PKT_ID