Enum libtegra::pinmux::PinFunction

pub enum PinFunction {
    Default,
    Aud,
    Bcl,
    Blink,
    Ccla,
    Cec,
    Cldvfs,
    Clk,
    Core,
    Cpu,
    Displaya,
    Displayb,
    Dmic1,
    Dmic2,
    Dmic3,
    Dp,
    Dtv,
    Extperiph3,
    I2C1,
    I2C2,
    I2C3,
    I2Cpmu,
    I2Cvi,
    I2S1,
    I2S2,
    I2S3,
    I2S4A,
    I2S4B,
    I2S5A,
    I2S5B,
    Iqc0,
    Iqc1,
    Jtag,
    Pe,
    Pe0,
    Pe1,
    Pmi,
    Pwm0,
    Pwm1,
    Pwm2,
    Pwm3,
    Qspi,
    Sata,
    Sdmmc1,
    Sdmmc3,
    Shutdown,
    Soc,
    Sor0,
    Sor1,
    Spdif,
    Spi1,
    Spi2,
    Spi3,
    Spi4,
    Sys,
    Touch,
    Uart,
    Uarta,
    Uartb,
    Uartc,
    Uartd,
    Usb,
    Vgp1,
    Vgp2,
    Vgp3,
    Vgp4,
    Vgp5,
    Vgp6,
    Vimclk,
    Vimclk2,
    Rsvd0,
    Rsvd1,
    Rsvd2,
    Rsvd3,
    Reserved,
}

Pin Functions that can be loaded onto a pin to control its behavior.

Note that most pins are actually predestined for a specific pin functions among reserved ones. The driver will make sure that no unsupported pin function can be set on the wrong pin.

Variants

Default

Aud

Bcl

Blink

Ccla

Cec

Cldvfs

Clk

Core

Cpu

Displaya

Displayb

Dmic1

Dmic2

Dmic3

Dp

Dtv

Extperiph3

I2C1

I2C2

I2C3

I2Cpmu

I2Cvi

I2S1

I2S2

I2S3

I2S4A

I2S4B

I2S5A

I2S5B

Iqc0

Iqc1

Jtag

Pe

Pe0

Pe1

Pmi

Pwm0

Pwm1

Pwm2

Pwm3

Qspi

Sata

Sdmmc1

Sdmmc3

Shutdown

Soc

Sor0

Sor1

Spdif

Spi1

Spi2

Spi3

Spi4

Sys

Touch

Uart

Uarta

Uartb

Uartc

Uartd

Usb

Vgp1

Vgp2

Vgp3

Vgp4

Vgp5

Vgp6

Vimclk

Vimclk2

Rsvd0

Rsvd1

Rsvd2

Rsvd3

Reserved