Enum libtegra::pinmux::PinGrP

pub enum PinGrP {
    Sdmmc1ClkPm0,
    Sdmmc1CmdPm1,
    Sdmmc1Dat3Pm2,
    Sdmmc1Dat2Pm3,
    Sdmmc1Dat1Pm4,
    Sdmmc1Dat0Pm5,
    Sdmmc3ClkPp0,
    Sdmmc3CmdPp1,
    Sdmmc3Dat0Pp5,
    Sdmmc3Dat1Pp4,
    Sdmmc3Dat2Pp3,
    Sdmmc3Dat3Pp2,
    PexL0RstNPa0,
    PexL0ClkreqNPa1,
    PexWakeNPa2,
    PexL1RstNPa3,
    PexL1ClkreqNPa4,
    SataLedActivePa5,
    Spi1MosiPc0,
    Spi1MisoPc1,
    Spi1SckPc2,
    Spi1Cs0Pc3,
    Spi1Cs1Pc4,
    Spi2MosiPb4,
    Spi2MisoPb5,
    Spi2SckPb6,
    Spi2Cs0Pb7,
    Spi2Cs1Pdd0,
    Spi4MosiPc7,
    Spi4MisoPd0,
    Spi4SckPc5,
    Spi4Cs0Pc6,
    QspiSckPee0,
    QspiCsNPee1,
    QspiIo0Pee2,
    QspiIo1Pee3,
    QspiIo2Pee4,
    QspiIo3Pee5,
    Dmic1ClkPe0,
    Dmic1DatPe1,
    Dmic2ClkPe2,
    Dmic2DatPe3,
    Dmic3ClkPe4,
    Dmic3DatPe5,
    Gen1I2CSclPj1,
    Gen1I2CSdaPj0,
    Gen2I2CSclPj2,
    Gen2I2CSdaPj3,
    Gen3I2CSclPf0,
    Gen3I2CSdaPf1,
    CamI2CSclPs2,
    CamI2CSdaPs3,
    PwrI2CSclPy3,
    PwrI2CSdaPy4,
    Uart1TxPu0,
    Uart1RxPu1,
    Uart1RtsPu2,
    Uart1CtsPu3,
    Uart2TxPg0,
    Uart2RxPg1,
    Uart2RtsPg2,
    Uart2CtsPg3,
    Uart3TxPd1,
    Uart3RxPd2,
    Uart3RtsPd3,
    Uart3CtsPd4,
    Uart4TxPi4,
    Uart4RxPi5,
    Uart4RtsPi6,
    Uart4CtsPi7,
    Dap1FsPb0,
    Dap1DinPb1,
    Dap1DoutPb2,
    Dap1SclkPb3,
    Dap2FsPaa0,
    Dap2DinPaa2,
    Dap2DoutPaa3,
    Dap2SclkPaa1,
    Dap4FsPj4,
    Dap4DinPj5,
    Dap4DoutPj6,
    Dap4SclkPj7,
    Cam1MclkPs0,
    Cam2MclkPs1,
    JtagRtck,
    Clk32KIn,
    Clk32KOutPy5,
    BattBcl,
    ClkReq,
    CpuPwrReq,
    PwrIntN,
    Shutdown,
    CorePwrReq,
    AudMclkPbb0,
    DvfsPwmPbb1,
    DvfsClkPbb2,
    GpioX1AudPbb3,
    GpioX3AudPbb4,
    Pcc7,
    HdmiCecPcc0,
    HdmiIntDpHpdPcc1,
    SpdifOutPcc2,
    SpdifInPcc3,
    UsbVbusEn0Pcc4,
    UsbVbusEn1Pcc5,
    DpHpd0Pcc6,
    WifiEnPh0,
    WifiRstPh1,
    WifiWakeApPh2,
    ApWakeBtPh3,
    BtRstPh4,
    BtWakeApPh5,
    ApWakeNfcPh7,
    NfcEnPi0,
    NfcIntPi1,
    GpsEnPi2,
    GpsRstPi3,
    CamRstPs4,
    CamAfEnPs5,
    CamFlashEnPs6,
    Cam1PwdnPs7,
    Cam2PwdnPt0,
    Cam1StrobePt1,
    LcdTePy2,
    LcdBlPwmPv0,
    LcdBlEnPv1,
    LcdRstPv2,
    LcdGpio1Pv3,
    LcdGpio2Pv4,
    ApReadyPv5,
    TouchRstPv6,
    TouchClkPv7,
    ModemWakeApPx0,
    TouchIntPx1,
    MotionIntPx2,
    AlsProxIntPx3,
    TempAlertPx4,
    ButtonPowerOnPx5,
    ButtonVolUpPx6,
    ButtonVolDownPx7,
    ButtonSlideSwPy0,
    ButtonHomePy1,
    Pa6,
    Pe6,
    Pe7,
    Ph6,
    Pk0,
    Pk1,
    Pk2,
    Pk3,
    Pk4,
    Pk5,
    Pk6,
    Pk7,
    Pl0,
    Pl1,
    Pz0,
    Pz1,
    Pz2,
    Pz3,
    Pz4,
    Pz5,
    Reserved,
}

Pin Groups on the Tegra X1 SoC that can be customized and variably configured.

Many drivers of the libtegra crate depend on proper Pin Multiplexing settings for the specific board before they can be used.

Variants

Sdmmc1ClkPm0

Sdmmc1CmdPm1

Sdmmc1Dat3Pm2

Sdmmc1Dat2Pm3

Sdmmc1Dat1Pm4

Sdmmc1Dat0Pm5

Sdmmc3ClkPp0

Sdmmc3CmdPp1

Sdmmc3Dat0Pp5

Sdmmc3Dat1Pp4

Sdmmc3Dat2Pp3

Sdmmc3Dat3Pp2

PexL0RstNPa0

PexL0ClkreqNPa1

PexWakeNPa2

PexL1RstNPa3

PexL1ClkreqNPa4

SataLedActivePa5

Spi1MosiPc0

Spi1MisoPc1

Spi1SckPc2

Spi1Cs0Pc3

Spi1Cs1Pc4

Spi2MosiPb4

Spi2MisoPb5

Spi2SckPb6

Spi2Cs0Pb7

Spi2Cs1Pdd0

Spi4MosiPc7

Spi4MisoPd0

Spi4SckPc5

Spi4Cs0Pc6

QspiSckPee0

QspiCsNPee1

QspiIo0Pee2

QspiIo1Pee3

QspiIo2Pee4

QspiIo3Pee5

Dmic1ClkPe0

Dmic1DatPe1

Dmic2ClkPe2

Dmic2DatPe3

Dmic3ClkPe4

Dmic3DatPe5

Gen1I2CSclPj1

Gen1I2CSdaPj0

Gen2I2CSclPj2

Gen2I2CSdaPj3

Gen3I2CSclPf0

Gen3I2CSdaPf1

CamI2CSclPs2

CamI2CSdaPs3

PwrI2CSclPy3

PwrI2CSdaPy4

Uart1TxPu0

Uart1RxPu1

Uart1RtsPu2

Uart1CtsPu3

Uart2TxPg0

Uart2RxPg1

Uart2RtsPg2

Uart2CtsPg3

Uart3TxPd1

Uart3RxPd2

Uart3RtsPd3

Uart3CtsPd4

Uart4TxPi4

Uart4RxPi5

Uart4RtsPi6

Uart4CtsPi7

Dap1FsPb0

Dap1DinPb1

Dap1DoutPb2

Dap1SclkPb3

Dap2FsPaa0

Dap2DinPaa2

Dap2DoutPaa3

Dap2SclkPaa1

Dap4FsPj4

Dap4DinPj5

Dap4DoutPj6

Dap4SclkPj7

Cam1MclkPs0

Cam2MclkPs1

JtagRtck

Clk32KIn

Clk32KOutPy5

BattBcl

ClkReq

CpuPwrReq

PwrIntN

Shutdown

CorePwrReq

AudMclkPbb0

DvfsPwmPbb1

DvfsClkPbb2

GpioX1AudPbb3

GpioX3AudPbb4

Pcc7

HdmiCecPcc0

HdmiIntDpHpdPcc1

SpdifOutPcc2

SpdifInPcc3

UsbVbusEn0Pcc4

UsbVbusEn1Pcc5

DpHpd0Pcc6

WifiEnPh0

WifiRstPh1

WifiWakeApPh2

ApWakeBtPh3

BtRstPh4

BtWakeApPh5

ApWakeNfcPh7

NfcEnPi0

NfcIntPi1

GpsEnPi2

GpsRstPi3

CamRstPs4

CamAfEnPs5

CamFlashEnPs6

Cam1PwdnPs7

Cam2PwdnPt0

Cam1StrobePt1

LcdTePy2

LcdBlPwmPv0

LcdBlEnPv1

LcdRstPv2

LcdGpio1Pv3

LcdGpio2Pv4

ApReadyPv5

TouchRstPv6

TouchClkPv7

ModemWakeApPx0

TouchIntPx1

MotionIntPx2

AlsProxIntPx3

TempAlertPx4

ButtonPowerOnPx5

ButtonVolUpPx6

ButtonVolDownPx7

ButtonSlideSwPy0

ButtonHomePy1

Pa6

Pe6

Pe7

Ph6

Pk0

Pk1

Pk2

Pk3

Pk4

Pk5

Pk6

Pk7

Pl0

Pl1

Pz0

Pz1

Pz2

Pz3

Pz4

Pz5

Reserved

Implementations

impl PinGrP

pub unsafe fn set_function(&self, function: PinFunction)

Configures a given Pin Function for this Pin Group.

Applicable to all pads, but only certain functions may be supported per pad.

Panics

Panics if the user attempts to set an unsupported pin function on this pad.

Safety

Playing around with Pin Multiplexing settings can irreparably damage your hardware, please make sure that you know exactly what you are doing before calling this function.

pub fn get_pull(&self) -> PinPull

Extracts the currently configured Pull resistor state from this Pin Group.

pub unsafe fn set_pull(&self, pull: PinPull)

Configures a given Pull resistor state for this Pin Group.

Applicable to all pads.

Safety

Playing around with Pin Multiplexing settings can irreparably damage your hardware, please make sure that you know exactly what you are doing before calling this function.

pub fn get_tristate(&self) -> PinTristate

Extracts the currently configured Tri-State from this Pin Group.

pub unsafe fn set_tristate(&self, tristate: PinTristate)

Configures a given Tri-State for this Pin Group.

Applicable to all pads.

Safety

Playing around with Pin Multiplexing settings can irreparably damage your hardware, please make sure that you know exactly what you are doing before calling this function.

pub fn get_park(&self) -> PinPark

Extracts the currently configured Parking state from this Pin Group.

pub unsafe fn set_park(&self, park: PinPark)

Sets the Parking state for this Pin Group.

Applicable to all pads except for a few ones in the AO region. This method does not validate the settings it is being called with, the user must ensure that they are calling this method correctly. See the advice below.

Safety

Playing around with Pin Multiplexing settings can irreparably damage your hardware, please make sure that you know exactly what you are doing before calling this function.

pub fn get_io(self) -> PinIo

Extracts the currently configured I/O state from this Pin Group.

pub unsafe fn set_io(&self, io: PinIo)

Sets a given I/O configuration for this Pin Group.

Applicable to all pads.

Safety

Playing around with Pin Multiplexing settings can irreparably damage your hardware, please make sure that you know exactly what you are doing before calling this function.

pub fn get_lock(&self) -> PinLock

Extracts the currently configured Lock control state from this Pin Group.

pub unsafe fn set_lock(&self, lock: PinLock)

Configures a given Lock control state for this Pin Group.

Applicable to all pads.

Safety

Playing around with Pin Multiplexing settings can irreparably damage your hardware, please make sure that you know exactly what you are doing before calling this function.

pub fn get_lpdr(&self) -> PinLpdr

Extracts the currently configured LPDR state from this Pin Group.

pub unsafe fn set_lpdr(&self, lpdr: PinLpdr)

Configures a given LPDR state for this Pin Group.

Applicable to ST and DD pads. This method does not validate the settings it is being called with, the user must ensure that they are calling this method correctly. See the advice below.

Safety

Playing around with Pin Multiplexing settings can irreparably damage your hardware, please make sure that you know exactly what you are doing before calling this function.

pub fn get_io_hv(&self) -> PinIoHv

Extracts the currently configured operation voltage state from this Pin Group.

pub unsafe fn set_io_hv(&self, hv: PinIoHv)

Configures the given operation voltage state for this Pin Group.

Applicable to DD pads. This method does not validate the settings it is being called with, the user must ensure that they are calling this method correctly. See the advice below.

Safety

Playing around with Pin Multiplexing settings can irreparably damage your hardware, please make sure that you know exactly what you are doing before calling this function.

pub fn get_od(&self) -> PinOd

Extracts the currently configured OD state from this Pin Group.

pub unsafe fn set_od(&self, od: PinOd)

Configures a given OD state for this Pin Group.

This field is marked and should never be set by user code for any pin.

Safety

Playing around with Pin Multiplexing settings can irreparably damage your hardware, please make sure that you know exactly what you are doing before calling this function.

pub fn get_schmt(&self) -> PinSchmt

Extracts the currently configured Schmitt state from this Pin Group.

pub unsafe fn set_schmt(&self, schmt: PinSchmt)

Configures Schmitt mode for this Pin Group.

Applicable to all pads.

Safety

Playing around with Pin Multiplexing settings can irreparably damage your hardware, please make sure that you know exactly what you are doing before calling this function.

pub fn get_drive(&self) -> PinDrive

Extracts the currently configured impedance state from this Pin Group.

pub unsafe fn set_drive(&self, drive: PinDrive)

Configures a given impedance drive for this Pin Group.

Applicable to CZ/LV_CZ pads. This method does not validate whether the drive setting is valid on the pin, the user must ensure that they are calling this method correctly. See the advice below.

Safety

Playing around with Pin Multiplexing settings can irreparably damage your hardware, please make sure that you know exactly what you are doing before calling this function.