Module libtegra::pwm

Driver for the Pulse Width Modulator Controller of the Tegra X1.

See Chapter 39 in the Tegra X1 Technical Reference Manual for details.

Description

The Pulse Width Modulator (PWM) Controller is a four-channel frequency divider whose pulse width varies. Each channel has a programmable frequency divider and a programmable pulse width generator.

Modules

PWM_CONTROLLER_PWM_CSR_0
Bitfields of the PWM_CONTROLLER_PWM_CSR_<x>_0 register.

Structs

PwmChannel
Representation of a Pulse Width Modulator channel.

Registers
Representation of the PWM Controller registers.

Constants

PWM_0_REGISTERS
A pointer to the PWM_0 register block that can be accessed by dereferencing it.

PWM_1_REGISTERS
A pointer to the PWM_1 register block that can be accessed by dereferencing it.

PWM_2_REGISTERS
A pointer to the PWM_2 register block that can be accessed by dereferencing it.

PWM_3_REGISTERS
A pointer to the PWM_3 register block that can be accessed by dereferencing it.