Struct libtegra::pwm::PwmChannel

pub struct PwmChannel { /* fields omitted */ }

Representation of a Pulse Width Modulator channel.

NOTE: It is expected that the PWM Clock is brought up before using any of the functionality of a channel.

Implementations

impl PwmChannel

pub const PWM_0: Self

Representation of the PWM 0 channel.

pub const PWM_1: Self

Representation of the PWM 1 channel.

pub const PWM_2: Self

Representation of the PWM 2 channel.

pub const PWM_3: Self

Representation of the PWM 3 channel.

impl PwmChannel

pub fn enable(&self)

Enables pulse generation mechanism of this channel.

pub fn disable(&self)

Disables pulse generation mechanism of this channel.

pub fn set_pulse_width(&self, duty: f32) -> Result<(), ()>

Configures the pulse width of the channel.

The argument is the desired duty cycle as a float value, representing a percentage ranging from 0.0 (0%) to 1.0 (100%).

pub fn get_duty(&self) -> f32

Returns the current duty cycle.

The returned duty cycle is a float value from 0.0 (0%) to 1.0 (100%).

pub fn get_max_duty(&self) -> f32

Returns the max duty cycle that is possible to set.