Module libtegra::se

Driver for the Tegra X1 Security Engine.

Description

The Security Engine is responsible for performing cryptographic operations in a hardware-based environment. Communication with hardware is done over a variety of supported interfaces, including memory and hash registers. It supports most modern cryptographic algorithms and is suited for secure operation as the engine has the ability to restrict access to it from ARM TrustZone clients.

Hardware Operations

As mentioned previously, the Security Engine performs hardware-based operations to process DMA buffers of data. These operations can be performed through the trigger_operation function.

As this is a low-level interface that should only be used if no other possibility exists, there are higher-level wrappers around commonly used operations available:

// TODO

Cryptographic APIs

The following cryptographic APIs are exposed by the Security Engine and can be used from TrustZone clients with no regrets:

Access Management

In secure systems, it is undesirable to have untrusted pieces of code access the Security Engine to perform arbitrary cryptographic operations. For that reason, the SE offers various features for locking its functionality down to Secure World clients which run in the ARM TrustZone. Great flexibility is given in how this is achieved: Full lockdown, lockdown per key, TZRAM lockdown, restriction of context save operations, full Security Engine shutdown. While not directly cryptographic, still an important aspect of retaining cryptographic security in applications.

SecurityEngine::lock

SecurityEngine::unlock

SecurityEngine::lock_per_key

SecurityEngine::lock_tzram

SecurityEngine::lock_context_save

RNG

The Security Engine implements a Random Number Generator which can be used to implement secure random data generation with a high entropy for applications.

SecurityEngine::initialize_rng

SecurityEngine::generate_random

SecurityEngine::set_random_key

SecurityEngine::generate_srk

AES

The AES APIs of the Security Engine expose the primitives to do cryptographic operations with various block and cipher modes. These operations source their keys from a secure hardware key table cache, whose contents can be individually manipulated. Further, access can be restricted under a specific set of circumstances to provide a secure operating environment without leaking keys to the outside.

SecurityEngine::fill_aes_keyslot

SecurityEngine::get_aes_key

SecurityEngine::clear_aes_keyslot

SecurityEngine::clear_aes_key_iv

SecurityEngine::set_encrypted_aes_key

SecurityEngine::aes_cmac

SecurityEngine::aes_ecb_encrypt

SecurityEngine::aes_ecb_decrypt

SecurityEngine::aes_cbc_encrypt

SecurityEngine::aes_cbc_decrypt

SecurityEngine::aes_ctr_encrypt

SecurityEngine::aes_ctr_decrypt

RSA

Similarly to the AES APIs, the Security Engine also features asymmetric encryptions using RSA where a user first feeds the desired exponent and modulus values for the encryption into keyslots and then triggers a modular exponentiation operation.

SecurityEngine::fill_rsa_keyslot

SecurityEngine::clear_rsa_keyslot

SecurityEngine::rsa_modular_exponentiate

Hashing

The Security Engine supports various hashing algorithms from the SHA1 and SHA2 family to calculate a fixed-size digest over a given buffer of data using the following methods.

SecurityEngine::calculate_sha1

SecurityEngine::calculate_sha224

SecurityEngine::calculate_sha256

SecurityEngine::calculate_sha384

SecurityEngine::calculate_sha512

Modules

SE_CONFIG_0

SE_CRYPTO_CONFIG_0

SE_CRYPTO_KEYTABLE_ADDR_0

SE_CRYPTO_KEYTABLE_DST_0

SE_CTX_SAVE_AUTO_0

SE_CTX_SAVE_CONFIG_0

SE_INT_ENABLE_0

SE_INT_STATUS_0

SE_OPERATION_0

SE_RNG_CONFIG_0

SE_RNG_SRC_CONFIG_0

SE_RSA_CONFIG_0

SE_RSA_KEYTABLE_ACCESS_0

SE_RSA_KEYTABLE_ADDR_0

SE_SE_SECURITY_0

SE_SHA_CONFIG_0

SE_STATUS_0

SE_TZRAM_OPERATION_0

SE_TZRAM_SECURITY_0

Structs

AddressInfo
Address information of a DMA buffer.

LinkedList
Representation of a Security Engine Linked List.

Registers
Representation of the Security Engine registers.

SecurityEngine
Representation of the Security Engine used for cryptographic operations.

Enums

AesMode
Representation of the different modes of operation for the AES algorithm for use in Security Engine operations.

OperationError
Enumeration of potential Security Engine errors that may occur during internal operations on it.

Constants

SE1_REGISTERS
A pointer to the SE1 register block that can be accessed by dereferencing it.

SE2_REGISTERS
A pointer to the SE2 register block that can be accessed by dereferencing it.

Functions

start_context_save_operation
Triggers a Security Engine operation that saves the SE context afterwards.

start_normal_operation
Triggers a regular Security Engine operation.

trigger_operation
Launches a cryptographic Security Engine operation, given two LinkedLists for I/O.