Module libtegra::spi

Driver for the Tegra X1 Serial Peripheral Interface Controller.

Modules

SPI_COMMAND2_0
Bitfields of the SPI_COMMAND2_0 register.

SPI_COMMAND_0
Bitfields of the SPI_COMMAND_0 register.

SPI_DMA_BLK_SIZE_0
Bitfields of the SPI_DMA_BLK_SIZE_0 register.

SPI_DMA_CTL_0
Bitfields of the SPI_DMA_CTL_0 register.

SPI_FIFO_STATUS_0
Bitfields of the SPI_FIFO_STATUS_0 register.

SPI_INTR_MASK_0
Bitfields of the SPI_INTR_MASK_0 register.

SPI_RX_DATA_0
Bitfields of the SPI_RX_DATA_0 register.

SPI_RX_FIFO_0
Bitfields of the SPI_RX_FIFO_0 register.

SPI_SPARE_CTLR
Bitfields of the SPI_SPARE_CTLR register.

SPI_TIMING_REG1_0
Bitfields of the SPI_TIMING_REG1_0 register.

SPI_TIMING_REG2_0
Bitfields of the SPI_TIMING_REG2_0 register.

SPI_TRANSFER_STATUS_0
Bitfields of the SPI_TRANSFER_STATUS_0 register.

SPI_TX_DATA_0
Bitfields of the SPI_TX_DATA_0 register.

SPI_TX_FIFO_0
Bitfields of the SPI_TX_FIFO_0 register.

Structs

Registers
Representation of the SPI registers.

Spi
Representation of an SPI.

Constants

QSPI_REGISTERS
A pointer to the QSPI register block that can be accessed by dereferencing it.

SPI_1_REGISTERS
A pointer to the SPI 2B-1 register block that can be accessed by dereferencing it.

SPI_2_REGISTERS
A pointer to the SPI 2B-2 register block that can be accessed by dereferencing it.

SPI_3_REGISTERS
A pointer to the SPI 2B-3 register block that can be accessed by dereferencing it.

SPI_4_REGISTERS
A pointer to the SPI 2B-4 register block that can be accessed by dereferencing it.

SPI_5_REGISTERS
A pointer to the SPI 2B-5 register block that can be accessed by dereferencing it.

SPI_6_REGISTERS
A pointer to the SPI 2B-6 register block that can be accessed by dereferencing it.