Module libtegra::spi::SPI_COMMAND_0

Bitfields of the SPI_COMMAND_0 register.

Modules

BIDIR
Bidirectional Transfer Control Bit.

BIT_LEN
Represents the number of bits transferred in either Packed or Unpacked Mode.

BOTH_EN_BIT
Both MISO and MOSI can also be used to transfer at the same time when this bit is set.

CS_POL_INACTIVE_0
The inactive value of the external device's CS value, which is connected to CS0.

CS_POL_INACTIVE_1
The inactive value of the external device's CS value, which is connected to CS1.

CS_POL_INACTIVE_2
The inactive value of the external device's CS value, which is connected to CS2.

CS_POL_INACTIVE_3
The inactive value of the external device's CS value, which is connected to CS3.

CS_SEL
In Master Mode, these bits are used to select a slave in the multi-slave environment.

CS_SW_HW
Software control of the SPI_CS signal in Master Mode.

CS_SW_VAL
CS signal value in Master Mode.

EN_LE_BIT
Whether Little Endian Bit should be enabled.

EN_LE_BYTE
Whether Little Endian Byte should be enabled.

IDLE_SDA
Inactive data signal format.

MODE
The SPI interface clock mode according to the device with which it is communicating.

MS
Master/Slave Mode select.

PACKED
Packed mode enable bit.

PIO
Starts the transfer.

RESERVED
Reserved for future use.

RX_EN
Receive enable.

TX_EN
Transmit enable.

Structs

Register

Constants

BIDIR

BIT_LEN

BOTH_EN_BIT

CS_POL_INACTIVE_0

CS_POL_INACTIVE_1

CS_POL_INACTIVE_2

CS_POL_INACTIVE_3

CS_SEL

CS_SW_HW

CS_SW_VAL

EN_LE_BIT

EN_LE_BYTE

IDLE_SDA

MODE

MS

PACKED

PIO

RESERVED

RX_EN

TX_EN