Module libtegra::spi::SPI_FIFO_STATUS_0

Bitfields of the SPI_FIFO_STATUS_0 register.

Modules

CS_INACTIVE
Whether CS is deasserted when SPI is in Slave Mode.

ERR
Will be set to 1 when errors occur.

FRAME_END
Status bit that is set during Slave Continuos Mode, in case of an Sclk mismatch.

RESERVED
Reserved for future use.

RX_FIFO_FLUSH
Whether the RX FIFO should be flushed.

RX_FIFO_FULL
RX FIFO Full Status.

RX_FIFO_FULL_COUNT
Indicates the number of slots in the receive FIFO remaining before the FIFO is full.

RX_FIFO_OVF
RX FIFO Overflow.

RX_FIFO_UNR
RX FIFO Underrun.

RX_FiFO_EMPTY
RX FIFO Empty Status.

TX_FIFO_EMPTY
TX FIFO Empty Status.

TX_FIFO_EMPTY_COUNT
Indicates the number of slots in the transmit FIFO remaining before the FIFO is full.

TX_FIFO_FLUSH
Whether the TX FIFO should be flushed.

TX_FIFO_FULL
TX FIFO Full Status.

TX_FIFO_OVF
TX FIFO Overflow.

TX_FIFO_UNR
TX FIFO Underrun.

Structs

Register

Constants

CS_INACTIVE

ERR

FRAME_END

RESERVED

RX_FIFO_FLUSH

RX_FIFO_FULL

RX_FIFO_FULL_COUNT

RX_FIFO_OVF

RX_FIFO_UNR

RX_FiFO_EMPTY

TX_FIFO_EMPTY

TX_FIFO_EMPTY_COUNT

TX_FIFO_FLUSH

TX_FIFO_FULL

TX_FIFO_OVF

TX_FIFO_UNR