Module libtegra::timer

Abstractions of the timer interfaces provided by the Tegra X1.

See Chapter 8 in the Tegra X1 Technical Reference Manual for details.

Overview

Name	Primary Use	Related Interrupt	Secure	Freq.

RTC	Wall Clock Timer	RTC	Pseudo	32kHz

TMR	NVIDIA Generic Timers	TMR9-0 TMR10-13	Cfg	1MHz OSC

WDT	Per CPU/COP	WDT_<>	Cfg	1Mhz

TSC	Reference for GT	N/A	Yes	OSC

GT	ARM CPU Generic Timers	PPIs*	Yes	TSC

Usage

RTC

The Real-Time Clock can be used for time, alarms, countdowns, etc.

use libtegra::timer::sleep;

sleep(5); // Delays execution for five seconds.

TMR

These timers can also be used for reading time, but more accurately.

use libtegra::timer::usleep;

usleep(5_000_000); // Delays execution for five seconds.

Modules

rtc
Interface to the Tegra X1 Real-Time Clock.

timerus
Abstractions over the Fixed Time Base Registers.

Functions

get_microseconds
Reads the current time in microseconds.

get_milliseconds
Reads the current time in milliseconds.

get_seconds
Reads the current time in seconds.

msleep
Sleeps for a given duration in milliseconds.

sleep
Sleeps for a given duration in seconds.

usleep
Sleeps for a given duration in microseconds.