Module libtegra::timer::rtc

Interface to the Tegra X1 Real-Time Clock.

See Chapter 12 in the Tegra X1 Technical Reference Manual for details.

Description

The Real-Time Clock (RTC) module maintains seconds and milliseconds counters, and five alarm registers. The RTC is in the 'always on' power domain, allowing for the counters to run and alarms to trigger when the system is in low-power state. If configured, interrupts triggered by the RTC can cause the system to wake up from a low-power state.

Modules

APBDEV_RTC_BUSY_0
Bitfields of the APBDEV_RTC_BUSY_0 register.

APBDEV_RTC_CONTROL_0
Bitfields of the APBDEV_RTC_CONTROL_0 register.

APBDEV_RTC_CORRECTION_FACTOR_0
Bitfields of the APBDEV_RTC_CORRECTION_FACTOR_0 register.

APBDEV_RTC_INTR_MASK_0
Bitfields of the APBDEV_RTC_INTR_MASK_0 register.

APBDEV_RTC_INTR_SET_0
Bitfields of the APBDEV_RTC_INTR_SET_0 register.

APBDEV_RTC_INTR_SOURCE_0
Bitfields of the APBDEV_RTC_INTR_SOURCE_0 register.

APBDEV_RTC_INTR_STATUS_0
Bitfields of the APBDEV_RTC_INTR_STATUS_0 register.

APBDEV_RTC_MILLI_SECONDS_0
Bitfields of the APBDEV_RTC_MILLI_SECONDS_0 register.

APBDEV_RTC_MILLI_SECONDS_ALARM_0
Bitfields of the APBDEV_RTC_MILLI_SECONDS_ALARM_0 register.

APBDEV_RTC_MILLI_SECONDS_COUNTDOWN_ALARM_0
Bitfields of the APBDEV_RTC_MILLI_SECONDS_COUNTDOWN_ALARM_0 register.

APBDEV_RTC_SECONDS_0
Bitfields of the APBDEV_RTC_SECONDS_0 register.

APBDEV_RTC_SECONDS_ALARM0_0
Bitfields of the APBDEV_RTC_SECONDS_ALARM0_0 register.

APBDEV_RTC_SECONDS_ALARM1_0
Bitfields of the APBDEV_RTC_SECONDS_ALARM1_0 register.

APBDEV_RTC_SECONDS_COUNTDOWN_ALARM_0
Bitfields of the APBDEV_RTC_SECONDS_COUNTDOWN_ALARM_0 register.

APBDEV_RTC_SHADOW_SECONDS_0
Bitfields of the APBDEV_RTC_SHADOW_SECONDS_0 register.

Structs

Registers
Representation of the Real-Time Clock registers.

Constants

REGISTERS
A pointer to the RTC register block that can be accessed by dereferencing it.