Module libtegra::timer::timerus

Abstractions over the Fixed Time Base Registers.

See Chapter 8.8 in the Tegra X1 Technical Reference Manual for details.

Description

The USEC_CFG/CNTR_1US registers provide a fixed time base (in microseconds) to be used by the rest of the system regardless of the clk_m frequency (i.e., 12 MHz or 38.4 MHz).

Modules

TIMERUS_CNTR_1US_0
Bitfields of the TIMERUS_USEC_CFG_0 register.

TIMERUS_CNTR_FREEZE_0
Bitfields of the TIMERUS_CNTR_FREEZE_0 register.

TIMERUS_USEC_CFG_0
Bitfields of the TIMERUS_USEC_CFG_0 register.

Structs

Registers
Representation of the Fixed Time Base registers.

Constants

REGISTERS
A pointer to the Fixed Time Base register block that can be accessed by dereferencing it.