Enum libtegra::tsec::FalconExceptionClause

#[repr(u8)]
pub enum FalconExceptionClause {
    Trap0,
    Trap1,
    Trap2,
    Trap3,
    InvalidOpcode,
    AuthenticationEntry,
    PageMiss,
    PageMultipleMiss,
    BreakpointHit,
}

Enumeration of potential Falcon processor exception clauses that may occur during code execution on the TSEC.

Variants

Trap0

The software trap 0 that may be triggered by the TRAP instruction.

Trap1

The software trap 1 that may be triggered by the TRAP instruction.

Trap2

The software trap 2 that may be triggered by the TRAP instruction.

Trap3

The software trap 3 that may be triggered by the TRAP instruction.

InvalidOpcode

A trap that is triggered when the Falcon encounters an opcode it cannot decode.

AuthenticationEntry

A so-called Secure Fault was indicated by jumping to a secure page, but the MAC verification for the code region failed, resulting in this exception being raised.

PageMiss

A page fault occurred because the TLB failed to provide mappings for a virtual address on lookup.

PageMultipleMiss

A page fault occurred because the TLB contained multiple mappings for a single virtual address on lookup.

BreakpointHit

A breakpoint, which was set through the integrated hardware debugging interface, was hit during code execution.