Struct libtegra::tsec::Tsec

pub struct Tsec { /* private fields */ }

Representation of the Tegra Security Co-Processor.

Implementations

impl Tsec

pub const A: Self = Tsec { registers: TSEC_A_REGISTERS, }

Representation of the TSEC-A instance.

pub const B: Self = Tsec { registers: TSEC_B_REGISTERS, }

Representation of the TSEC-B instance.

impl Tsec

pub fn init(&self)

Initializes the TSEC for use.

NOTE: This method must be called once before the TSEC is usable. Otherwise, the SoC will hang itself whenever the device is accessed.

pub fn finalize(&self)

Shuts the TSEC down and makes it inaccessible.

pub fn load_firmware(&self, firmware: &[u8]) -> Result<(), FalconError>

Loads Falcon microcode into the processor memory.

This method utilizes the Falcon DMA engine to load the given firmware into the code segment, starting from physical and virtual address 0 and must be separately executed using Tsec::boot.

NOTE: The firmware buffer is expected to be aligned correctly to the boundaries of 0x100 byte pages in order to be uploaded.

pub unsafe fn boot(
&self,
boot_vector: u32,
mailbox0: &mut u32,
mailbox1: &mut u32
) -> Result<(), FalconError>

Boots the Falcon from the specified boot vector.

The firmware must have been loaded into the Falcon in advance, either by calling Tsec::load_firmware or doing the necessary transfers manually. The boot vector then specifies from where code should be executed, most code blobs presumably expect 0 to be passed, and the CPU boots up.

There is also support for both shared mailboxes which act as scratch registers to share data between the Falcon and the host system. Through the respective arguments, mailboxes can be filled with supplied values and at the end of execution, the variables will be overridden with the final state of the TSEC mailboxes (e.g. to check result codes on the host processor).

Safety

This method is considered unsafe because code execution on the TSEC can fail for malformed or misaligned blobs or through code fucking up internal state.

pub fn dump_dmem(&self, output: &mut [u32; 4096])

Dumps the DMEM of the Falcon engine into the supplied buffer.

This function is useful for debugging purposes and to examine how firmware interacts with the data memory.

The DMEM size of TSEC is hardcoded to 0x4000 bytes. For other Falcon engines, the size of DMEM in words should be determined by reading (FALCON_HWCFG >> 9 & 0x1F) << 6.

NOTE: This is only usable while TSEC is in No Secure mode context.