Module libtegra::uart::UART_LCR_0

Bitfields of the UART_LCR_0 register.

Modules

DLAB
Whether the Divisor Latch Access Bit should be enabled.

EVEN
Whether the even parity format should be used for number representation.

PAR
Whether parity should be sent or not.

SET_B
Whether a BREAK condition should be set.

SET_P
Whether parity should be set (forced) to the value in LCR.

STOP
Whether 2 stop bits should be transmitted instead of 1.

WD_SIZE
The Word Length size.

Structs

Register

Constants

DLAB

EVEN

PAR

SET_B

SET_P

STOP

WD_SIZE