Module libtegra::uart::UART_LSR_0

Bitfields of the UART_LSR_0 register.

Modules

BRK
Denotes that a BREAK condition was detected on the line, if set to 1.

FERR
Denotes a Framing Error, if set to 1.

FIFOE
Denotes a Receive FIFO error, if set to 1.

OVRF
Denotes a Receiver Overrun Error, if set to 1.

PERR
Denotes a Parity Error, if set to 1.

RDR
Denotes that Receiver Data are in FIFO, if set to 1.

RX_FIFO_EMPTY
Whether the RX FIFO is empty.

THRE
Denotes that the Transmit Holding Register is empty, if set to 1.

TMTY
Denotes a Transmit Shift Register empty status, if set to 1.

TX_FIFO_FULL
Whether the Transmitter FIFO is full.

Structs

Register

Constants

BRK

FERR

FIFOE

OVRF

PERR

RDR

RX_FIFO_EMPTY

THRE

TMTY

TX_FIFO_FULL