Module libtegra::uart::UART_VENDOR_STATUS_0_0

Bitfields of the UART_VENDOR_STATUS_0_0 register.

Modules

RX_FIFO_COUNTER
The entry in this field reflects the number of current entries in the RX FIFO.

RX_UNDERRUN
This bit is set to 1 when a read is issued to an empty FIFO and gets cleared on register read (sticky bit until read).

TX_FIFO_COUNTER
The entry in this field reflects the number of current entries in the TX FIFO.

TX_OVERRUN
This bit is set to 1 when write data is issued to the TX FIFO when it is already full and gets cleared on register read (sticky bit until read).

UART_RX_IDLE
This bit is set to 1 when the RX path is IDLE.

UART_TX_IDLE
This bit is set to 1 when the TX path is IDLE.

Structs

Register

Constants

RX_FIFO_COUNTER

RX_UNDERRUN

TX_FIFO_COUNTER

TX_OVERRUN

UART_RX_IDLE

UART_TX_IDLE