Module libtegra::vic::NV_PVIC_FALCON_CPUCTL

Bitfields of the NV_PVIC_FALCON_CPUCTL register.

Modules

NV_PVIC_FALCON_CPUCTL_HALTED

Indicates whether the CPU is currently in the halted state. Falcon can only exit this state when a 1 is written to the STARTCPU bit.

NV_PVIC_FALCON_CPUCTL_HRESET

Set to apply hard reset. This bit will auto-clear.

NV_PVIC_FALCON_CPUCTL_IINVAL

Set this bit to mark all blocks in IMEM except block 0 as INVALID. This bit will auto-clear.

NV_PVIC_FALCON_CPUCTL_SRESET

Set to apply soft reset. This bit will auto-clear.

NV_PVIC_FALCON_CPUCTL_STARTCPU

Set this bit to start CPU execution while in a halted state. If a start request is still pending, clearing this bit will cancel the start request. Writing any value has no effect while the CPU is running.

NV_PVIC_FALCON_CPUCTL_STOPPED

Indicates whether the CPU is currently in the stopped state. Falcon exits this state if a 1 is written to the STARTCPU bit or if an interrupt arrives at one of its 2 inputs and the corresponding IE bit in CSW is set.

Structs

Register

Constants

NV_PVIC_FALCON_CPUCTL_HALTED

NV_PVIC_FALCON_CPUCTL_HRESET

NV_PVIC_FALCON_CPUCTL_IINVAL

NV_PVIC_FALCON_CPUCTL_SRESET

NV_PVIC_FALCON_CPUCTL_STARTCPU

NV_PVIC_FALCON_CPUCTL_STOPPED