Crate libtegra

Re-exports

pub extern crate cortex_a;

Modules

apb

car
Abstractions over the Tegra X1 Clock and Reset Controller functionality.

flow
Driver for the Flow Controller of the Tegra X1.

gpio
Driver for the Tegra X1 General-Purpose Input/Output controller.

i2c
Driver for the Tegra X1 Inter-Integrated Circuit master and slave controllers.

kfuse
Driver for the Tegra X1 for the HDCP KFUSE Controller.

memory_map
Unified Tegra X1 Memory Address Map constants.

pinmux
Driver for Tegra X1 Multi-Purpose Pins and Pin Multiplexing.

pmc
Driver for the Tegra X1 Power Management Controller.

pwm
Driver for the PWM Controller of the Tegra X1.

spi
Driver for the Tegra X1 Serial Peripheral Interface Controller.

timer
Abstractions of the timer interfaces provided by the Tegra X1.

tsec
Driver for interfacing with the Tegra Security Co-Processor (TSEC).

uart
Driver for the Tegra X1 Universal Asynchronous Receiver/Transmitter Controller.

Macros

make_gpio
A macro to facilitate the creation of a GPIO given port and pin.