Module libtegra::actmon::ACTMON_APB_CTRL_SAPB_0::MONITOR_COND_SAPB

Selection criteria on parent module for the activity signal.

Only used for APB/AHB monitor.

Enums

Value

Selection criteria on parent module for the activity signal.

Constants

All

Apb2Jtag

Ape

Atomics

CLEAR

CSite

Cec

Dds

Dp2

Dvfs

Emc0

Emc1

EmcB

Fuse

Hda

KFuse

La

Mc0

Mc1

McB

MipiCal

Misc

PinmuxAux

Pmc

Pwm

Rtc

SET

SUxbPadctl

Sata

SataAux

Sdmmc1

Sdmmc2

Sdmmc3

Sdmmc4

Se

SecureRegs

SocTherm

Stm

SysCtr0

SysCtr1

XUsbDev

XUsbHost