Module libtegra::actmon::ACTMON_CTRL_0

Bitfields of the ACTMON_CPU_CTRL_0 register.

Modules

AHB_MONITOR_COND

Selection criteria on parent module for the activity signal.

AHB_SAMPLE_COND

Selection criteria on parent module for type of pulse signal.

APB_MONITOR_COND

Selection criteria on parent module for the activity signal.

APB_SAMPLE_COND

Selection criteria on parent module for type of pulse signal.

AT_END_EN

Enable interrupt at the end of sample period.

AVG_ABOVE_WMARK_EN

Enable interrupt when AVG value is above its upper watermark value.

AVG_BELOW_WMARK_EN

Enable interrupt when AVG value is below its lower watermark value.

CONSECUTIVE_ABOVE_WMARK_EN

Enable interrupt when consecutive CONSECUTIVE_UPPER_NUM upper watermark breaches are detected.

CONSECUTIVE_ABOVE_WMARK_NUM

Number (N+1) of consecutive upper watermark breaches that need to occur to raise an interrupt.

CONSECUTIVE_BELOW_WMARK_EN

Enable interrupt when consecutive CONSECUTIVE_LOWER_NUM lower watermark braches are detected.

CONSECUTIVE_BELOW_WMARK_NUM

Number (N+1) of consecutive lower watermark breaches that need to occur to raise an interrupt.

ENB

Enable Monitor. Set by software to enable sampling. Cleared in one of the following ways: (a) When software intends to stop the monitor, it can do so by clearing this field, (b) when the sampling period expires (and we are not in periodic mode).

ENB_PERIODIC

Enable periodic mode. Sample for one sample period or periodic sampling.

K_VAL

Variable for IIR filter. Default is 6, which translates to 2^(K+1) = 128.

WHEN_OVERFLOW_EN

Enable interrupt for number of consecutive lower watermark breaches that need to occur to raise an interrupt.

Structs

Register

Constants

AHB_MONITOR_COND

AHB_SAMPLE_COND

APB_MONITOR_COND

APB_SAMPLE_COND

AT_END_EN

AVG_ABOVE_WMARK_EN

AVG_BELOW_WMARK_EN

CONSECUTIVE_ABOVE_WMARK_EN

CONSECUTIVE_ABOVE_WMARK_NUM

CONSECUTIVE_BELOW_WMARK_EN

CONSECUTIVE_BELOW_WMARK_NUM

ENB

ENB_PERIODIC

K_VAL

WHEN_OVERFLOW_EN