Module libtegra::actmon::ACTMON_HISTOGRAM_CONFIG_0

Bitfields of the ACTMON_HISTOGRAM_CONFIG_0 register.

Modules

ACTIVE

Enable histogram recording.

LINEAR_MODE

Whether linear mode should be used to expand buckets.

NO_UNDERFLOW_BUCKET

Whether the underflow bucket should be incremented.

SHIFT

Scaling factor for the idle counter before the value is used to update histogram bucket.

SOURCE

STALL_ON_SINGLE_SATURATE

Whether buckets should be incremented when other buckets have saturated.

Structs

Register

Constants

ACTIVE

LINEAR_MODE

NO_UNDERFLOW_BUCKET

SHIFT

SOURCE

STALL_ON_SINGLE_SATURATE