Module libtegra::actmon::ACTMON_INTR_STATUS_0

Bitfields of the ACTMON_CPU_INTR_STATUS_0 register.

Modules

AT_END

Assert at the end of sampling period, if interrupt at the end of sample period is enabled.

AVG_ABOVE_WMARK

Assert at the end of sampling period, if AVG count value crosses upper AVG watermark value.

AVG_BELOW_WMARK

Assert at the end of sampling period, if AVG count value crosses lower AVG watermark value.

CONSECUTIVE_LOWER

Assert at the end of sampling period, if count value crosses lower watermark value consecutively for the number of times specified in CONSECUTIVE_LOWER_NUM.

CONSECUTIVE_UPPER

Assert at the end of sampling period, if count value crosses upper watermark value consecutively for the number of times specified in CONSECUTIVE_UPPER_NUM.

WHEN_OVERFLOW

Assert at the end of sampling period, if there is an overflow.

Structs

Register

Constants

AT_END

AVG_ABOVE_WMARK

AVG_BELOW_WMARK

CONSECUTIVE_LOWER

CONSECUTIVE_UPPER

WHEN_OVERFLOW