Struct libtegra::apb::dma::Channel

pub struct Channel { /* private fields */ }

Representation of an APB DMA Channel.

Channels are used for data transfers over DMA by the DMA Controller and need to be acquired and released when using them to avoid data races.

NOTE: Instances of this struct should never be created manually. Refer to the public constants this struct holds, which represent the channels 0 through 31.

Implementations

impl Channel

pub const CH0: Self = Channel { registers: CHANNEL_0, claimed: false, }

Representation of the DMA Channel 0.

pub const CH1: Self = Channel { registers: CHANNEL_1, claimed: false, }

Representation of the DMA Channel 1.

pub const CH2: Self = Channel { registers: CHANNEL_2, claimed: false, }

Representation of the DMA Channel 2.

pub const CH3: Self = Channel { registers: CHANNEL_3, claimed: false, }

Representation of the DMA Channel 3.

pub const CH4: Self = Channel { registers: CHANNEL_4, claimed: false, }

Representation of the DMA Channel 4.

pub const CH5: Self = Channel { registers: CHANNEL_5, claimed: false, }

Representation of the DMA Channel 5.

pub const CH6: Self = Channel { registers: CHANNEL_6, claimed: false, }

Representation of the DMA Channel 6.

pub const CH7: Self = Channel { registers: CHANNEL_7, claimed: false, }

Representation of the DMA Channel 7.

pub const CH8: Self = Channel { registers: CHANNEL_8, claimed: false, }

Representation of the DMA Channel 8.

pub const CH9: Self = Channel { registers: CHANNEL_9, claimed: false, }

Representation of the DMA Channel 9.

pub const CH10: Self = Channel { registers: CHANNEL_10, claimed: false, }

Representation of the DMA Channel 10.

pub const CH11: Self = Channel { registers: CHANNEL_11, claimed: false, }

Representation of the DMA Channel 11.

pub const CH12: Self = Channel { registers: CHANNEL_12, claimed: false, }

Representation of the DMA Channel 12.

pub const CH13: Self = Channel { registers: CHANNEL_13, claimed: false, }

Representation of the DMA Channel 13.

pub const CH14: Self = Channel { registers: CHANNEL_14, claimed: false, }

Representation of the DMA Channel 14.

pub const CH15: Self = Channel { registers: CHANNEL_15, claimed: false, }

Representation of the DMA Channel 15.

pub const CH16: Self = Channel { registers: CHANNEL_16, claimed: false, }

Representation of the DMA Channel 16.

pub const CH17: Self = Channel { registers: CHANNEL_17, claimed: false, }

Representation of the DMA Channel 17.

pub const CH18: Self = Channel { registers: CHANNEL_18, claimed: false, }

Representation of the DMA Channel 18.

pub const CH19: Self = Channel { registers: CHANNEL_19, claimed: false, }

Representation of the DMA Channel 19.

pub const CH20: Self = Channel { registers: CHANNEL_20, claimed: false, }

Representation of the DMA Channel 20.

pub const CH21: Self = Channel { registers: CHANNEL_21, claimed: false, }

Representation of the DMA Channel 21.

pub const CH22: Self = Channel { registers: CHANNEL_22, claimed: false, }

Representation of the DMA Channel 22.

pub const CH23: Self = Channel { registers: CHANNEL_23, claimed: false, }

Representation of the DMA Channel 23.

pub const CH24: Self = Channel { registers: CHANNEL_24, claimed: false, }

Representation of the DMA Channel 24.

pub const CH25: Self = Channel { registers: CHANNEL_25, claimed: false, }

Representation of the DMA Channel 25.

pub const CH26: Self = Channel { registers: CHANNEL_26, claimed: false, }

Representation of the DMA Channel 26.

pub const CH27: Self = Channel { registers: CHANNEL_27, claimed: false, }

Representation of the DMA Channel 27.

pub const CH28: Self = Channel { registers: CHANNEL_28, claimed: false, }

Representation of the DMA Channel 28.

pub const CH29: Self = Channel { registers: CHANNEL_29, claimed: false, }

Representation of the DMA Channel 29.

pub const CH30: Self = Channel { registers: CHANNEL_30, claimed: false, }

Representation of the DMA Channel 30.

pub const CH31: Self = Channel { registers: CHANNEL_31, claimed: false, }

Representation of the DMA Channel 31.

impl Channel

pub fn acquire(&mut self) -> Result<(), ()>

Acquires the selected channel.

Returns an error if the channel is already claimed at the time of calling the method. When the channel is not needed anymore, it must be freed with Channel::release.

pub fn release(&mut self) -> Result<(), ()>

Releases the selected channel after it is not needed anymore.

Returns an error if the channel is not claimed at the time of calling the method.

pub fn is_acquired(&self) -> bool

Indicates whether the channel is currently claimed.

pub fn is_free(&self) -> bool

Indicates whether the selected channel is currently unused.

This is the counterpart to Channel::is_acquired.

pub fn start(&self)

Starts a DMA transfer in the selected channel.

pub fn finish(&self)

Finishes a DMA transfer in the selected channel.

pub fn is_busy(&self) -> bool

Indicates whether the channel is currently busy doing transfers.

NOTE: Channels need to be acquired before being able to transfer data. If this method returns true, Channel::is_acquired is guaranteed to return true as well.

pub fn is_idle(&self) -> bool

Indicates whether the channel is in idle state and ready for transfers.

This is the counterpart to Channel::is_busy.

pub fn query(
&self,
slave: u32,
ahb_address: u32,
apb_address: u32,
size: u32
) -> Result<(), ()>

Prepares data to be queried through the selected channel over DMA.

This function doesn’t actually trigger transfers, it only prepares the data output buffer and configures the channel. The actual transfer process can be initiated through Channel::start and terminated through Channel::finish. For transfer status details, see Channel::is_busy.

pub fn write(
&self,
slave: u32,
ahb_address: u32,
apb_address: u32,
size: u32
) -> Result<(), ()>

Prepares data to be written through the selected channel over DMA.

This method doesn’t actually trigger transfers, it only loads in the data and configures the channel. The actual transfer process can be initiated through Channel::start and terminated through Channel::finish. For transfer status details, see Channel::is_busy.