Module libtegra::arm

Utilities related to the ARM processors found on Tegra X1 SoC.

This covers both, the ARM7TDMI used as the [bpmp] and the Cortex-A53/A57 processors forming the main CPU Complex (CCPLEX).

Modules

gic

Implementation of the ARM Generic Interrupt Controller v2.

Functions

align_down

Aligns address down to the boundary denoted by align.

align_up

Aligns address up to the boundary denoted by align.