Module libtegra::car

Abstractions over the Tegra X1 Clock and Reset Controller functionality.

See Chapter 5 in the Tegra X1 Technical Reference Manual for details.

Description

The Clock and Reset (CAR) block contains all the logic needed to control most of the clocks and resets to the Tegra X1 device. It takes care of most clock source programming and most of the clock dividers.

Clocks

Clocks provide the logic to control the operation frequency of hardware blocks, which include certain SoC devices, such as system buses, timers, the RTC, GPIO and CoreSight. The Clock structure provides an abstraction layer to control the power-on (POR) and reset states of each block individually.

Instances of Clock should never be created by the user, as malformed configurations bring a certain risk of damaging the hardware. That’s why the structure holds globally-accessible constant instances of itself, which represent most of the known device clocks without non-standard programming guidelines.

Structs

Clock

Representation of a device clock.

Registers

Representation of the CAR registers.

Constants

CLK_H_AHBDMA

CLK_H_APBDMA

CLK_H_BSEA

CLK_H_BSEV

CLK_H_CSI

CLK_H_DSI

CLK_H_EMC

CLK_H_FUSE

CLK_H_HDMI

CLK_H_HSI

CLK_H_I2C2

CLK_H_I2C5

CLK_H_JTAG2TBC

CLK_H_KBC

CLK_H_KFUSE

CLK_H_MEM

CLK_H_MIPI_CAL

CLK_H_MPE

CLK_H_PMC

CLK_H_SNOR

CLK_H_SPI1

CLK_H_SPI2

CLK_H_SPI3

CLK_H_STAT_MON

CLK_H_UARTC

CLK_H_USB2

CLK_H_USB3

CLK_H_VDE

CLK_H_XIO

CLK_L_2D

CLK_L_3D

CLK_L_BPMP

CLK_L_BPMP_CACHE_CTRL

CLK_L_CPU

CLK_L_DISP1

CLK_L_DISP2

CLK_L_EPP

CLK_L_GPIO

CLK_L_HOST1X

CLK_L_I2C1

CLK_L_I2S1

CLK_L_I2S2

CLK_L_I2S3

CLK_L_ISP

CLK_L_ISPB

CLK_L_NDFLASH

CLK_L_PWM

CLK_L_RTC

CLK_L_SDMMC1

CLK_L_SDMMC2

CLK_L_SDMMC4

CLK_L_SPDIF

CLK_L_SYS

CLK_L_TMR

CLK_L_TWC

CLK_L_UARTA

CLK_L_UARTB

CLK_L_USBD

CLK_L_VCP

CLK_L_VI

CLK_NO_SOURCE

CLK_RST_CONTROLLER_CLK_OUT_ENB_H

CLK_RST_CONTROLLER_CLK_OUT_ENB_L

CLK_RST_CONTROLLER_CLK_OUT_ENB_U

CLK_RST_CONTROLLER_CLK_OUT_ENB_V

CLK_RST_CONTROLLER_CLK_OUT_ENB_W

CLK_RST_CONTROLLER_CLK_OUT_ENB_X

CLK_RST_CONTROLLER_CLK_OUT_ENB_Y

CLK_RST_CONTROLLER_CLK_SOURCE_CSITE

CLK_RST_CONTROLLER_CLK_SOURCE_HOST1X

CLK_RST_CONTROLLER_CLK_SOURCE_I2C_1

CLK_RST_CONTROLLER_CLK_SOURCE_I2C_2

CLK_RST_CONTROLLER_CLK_SOURCE_I2C_3

CLK_RST_CONTROLLER_CLK_SOURCE_I2C_4

CLK_RST_CONTROLLER_CLK_SOURCE_I2C_5

CLK_RST_CONTROLLER_CLK_SOURCE_I2C_6

CLK_RST_CONTROLLER_CLK_SOURCE_PWM

CLK_RST_CONTROLLER_CLK_SOURCE_SE

CLK_RST_CONTROLLER_CLK_SOURCE_SOR1

CLK_RST_CONTROLLER_CLK_SOURCE_TSEC

CLK_RST_CONTROLLER_CLK_SOURCE_TSECB

CLK_RST_CONTROLLER_CLK_SOURCE_UART_A

CLK_RST_CONTROLLER_CLK_SOURCE_UART_APE

CLK_RST_CONTROLLER_CLK_SOURCE_UART_B

CLK_RST_CONTROLLER_CLK_SOURCE_UART_C

CLK_RST_CONTROLLER_CLK_SOURCE_UART_D

CLK_RST_CONTROLLER_RST_DEVICES_H

CLK_RST_CONTROLLER_RST_DEVICES_L

CLK_RST_CONTROLLER_RST_DEVICES_U

CLK_RST_CONTROLLER_RST_DEVICES_V

CLK_RST_CONTROLLER_RST_DEVICES_W

CLK_RST_CONTROLLER_RST_DEVICES_X

CLK_RST_CONTROLLER_RST_DEVICES_Y

CLK_U_AFI

CLK_U_BPMPUCQ

CLK_U_BPMP_CACHE_RAM

CLK_U_CLK_M_DOUBLER

CLK_U_CSITE

CLK_U_DEV1_OUT

CLK_U_DEV2_OUT

CLK_U_DSIB

CLK_U_DTV

CLK_U_I2C3

CLK_U_I2C_SLOW

CLK_U_IRAMA

CLK_U_IRAMB

CLK_U_IRAMC

CLK_U_IRAMD

CLK_U_LA

CLK_U_MSENC

CLK_U_NAND_SPEED

CLK_U_PCIE

CLK_U_PCIEXCLK

CLK_U_SDMMC3

CLK_U_SOC_THERM

CLK_U_SPI4

CLK_U_SUS_OUT

CLK_U_TRACECLKIN

CLK_U_TSEC

CLK_U_UARTD

CLK_U_UARTE

CLK_U_UNUSED

CLK_U_XUSB_DEV

CLK_U_XUSB_HOST

CLK_V_3D2

CLK_V_ACTMON

CLK_V_AHUB

CLK_V_APB2APE

CLK_V_ATOMICS

CLK_V_CPUG

CLK_V_CPULP

CLK_V_DAM0

CLK_V_DAM1

CLK_V_DAM2

CLK_V_EXTPERIPH1

CLK_V_EXTPERIPH2

CLK_V_EXTPERIPH3

CLK_V_HDA

CLK_V_HDA2CODEC_2X

CLK_V_I2C4

CLK_V_I2S4

CLK_V_I2S5

CLK_V_MSELECT

CLK_V_SATA

CLK_V_SATA_OOB

CLK_V_SE

CLK_V_SPDIF_DOUBLER

CLK_V_SPI5

CLK_V_SPI6

CLK_V_TSENSOR

CLK_V_TZRAM

CLK_W_ADX0

CLK_W_AMX0

CLK_W_CEC

CLK_W_CILAB

CLK_W_CILCD

CLK_W_CILEF

CLK_W_DDS

CLK_W_DP2

CLK_W_DSIA_LP

CLK_W_DSIB_LP

CLK_W_DVFS

CLK_W_EMC_IOBIST

CLK_W_EMC_LATENCY

CLK_W_ENTROPY

CLK_W_HDA2HDMICODEC

CLK_W_HDMI_IOBIST

CLK_W_MC1

CLK_W_MIPI_IOBIST

CLK_W_PCIE2_IOBIST

CLK_W_PCIERX0

CLK_W_PCIERX1

CLK_W_PCIERX2

CLK_W_PCIERX3

CLK_W_PCIERX4

CLK_W_PCIERX5

CLK_W_RESERVED0

CLK_W_SATA_IOBIST

CLK_W_XUSB

CLK_W_XUSB_PADCTL

CLK_W_XUSB_SS

CLK_X_ADX1

CLK_X_CAM_MCLK

CLK_X_CAM_MCLK2

CLK_X_DBGAPB

CLK_X_DMIC1

CLK_X_DMIC2

CLK_X_DPAUX

CLK_X_EMC_DLL

CLK_X_ETR

CLK_X_GPU

CLK_X_HDMI_AUDIO

CLK_X_HPLL_ADSP

CLK_X_I2C6

CLK_X_MC_BBC

CLK_X_MC_CAPA

CLK_X_MC_CBPA

CLK_X_MC_CPU

CLK_X_MIPIBIF

CLK_X_PLLA_ADSP

CLK_X_PLLG_REF

CLK_X_PLLP_ADSP

CLK_X_SOR0

CLK_X_SOR1

CLK_X_SPARE

CLK_X_UART_FST_MIPI_CAL

CLK_X_VIC

CLK_X_VIM2_CLK

CLK_Y_ADSP

CLK_Y_ADSPDBG

CLK_Y_ADSPINTF

CLK_Y_ADSPNEON

CLK_Y_ADSPPERIPH

CLK_Y_ADSPSCU

CLK_Y_ADSPWDT

CLK_Y_APE

CLK_Y_AXIAP

CLK_Y_DMIC3

CLK_Y_DPAUX1

CLK_Y_HSIC_TRK

CLK_Y_IQC1

CLK_Y_IQC2

CLK_Y_MAUD

CLK_Y_MC_CCPA

CLK_Y_MC_CDPA

CLK_Y_NVDEC

CLK_Y_NVENC

CLK_Y_NVJPG

CLK_Y_PEX_USB_UPHY

CLK_Y_PLLP_OUT_CPU

CLK_Y_QSPI

CLK_Y_SATA_USB_UPHY

CLK_Y_SDMMC_LEGACY_TM

CLK_Y_SOR_SAFE

CLK_Y_SPARE1

CLK_Y_TSECB

CLK_Y_UARTAPE

CLK_Y_USB2_TRK

CLK_Y_VI_I2C

REGISTERS

A pointer to the CAR register block that can be accessed by dereferencing it.