Module libtegra::dsi::DSI_DSI_CONTROL_0

Bitfields of the DSI_DSI_CONTROL_0 register.

Modules

DFMT_16BPP_SWAP_EN

DSI specification supports different bit orderings (only 16 bpp) in command mode.

DSI_DATA_FORMAT

Pixel Data format transmitted. Only this information for constructing RGB data packets in hardware.

DSI_DBG_ENABLE

Control signal to turn off clock monitoring when enabled for debug, on every DSI byte clock debug signal toggle.

DSI_HOST_ENABLE

Host DSI Interface Enable.

DSI_HS_CLK_CTRL

Control for the HS clock lane.

DSI_NUM_DATA_LANES

Number of D-PHY data lanes used by Display for HS transmission.

DSI_VID_ENABLE

Video DSI Interface Enable.

DSI_VID_SOURCE

Source of video pixels.

DSI_VIRTUAL_CHANNEL

Virtual channel ID. Sent as part of the packet header and used to distinguish multiple displays.

VID_DCS_ENABLE

Allows insertion of DCS commands during Display Controller generated packets.

VID_TX_TRIG_SRC

Controls the source of the trigger to start sending packets.

Structs

Register

Constants

DFMT_16BPP_SWAP_EN

DSI_DATA_FORMAT

DSI_DBG_ENABLE

DSI_HOST_ENABLE

DSI_HS_CLK_CTRL

DSI_NUM_DATA_LANES

DSI_VID_ENABLE

DSI_VID_SOURCE

DSI_VIRTUAL_CHANNEL

VID_DCS_ENABLE

VID_TX_TRIG_SRC