Module libtegra::dsi::DSI_HOST_DSI_CONTROL_0

Bitfields of the DSI_HOST_DSI_CONTROL_0 register.

Modules

CRC_RESET

Whether the Verification CRC should be reset to 0xFFFF_FFFF.

CS_ENABLE

Enable hardware Check Sum (CS) for Host packets.

DSI_HIGH_SPEED_TRANS

DSI high speed transmission of packets.

DSI_PHY_CLK_DIV

Physical clock divider value for byte clock.

DSI_ULTRA_LOW_POWER

Ultra Low Power mode.

ECC_ENABLE

Enable hardware Error Code Correction (ECC) for Host packets.

FIFO_STAT_RESET

Clears FIFO underflow/overflow flags.

HOST_TX_TRIG_SRC

Controls the source of the trigger to start sending packets.

IMM_BTA

Generate BTA immediately, e.g. for Tearing Effect reporting.

PERIPH_RESET

Initiate an Escape Mode Peripheral Reset.

PKT_BTA

Generate BTA at the end of Host packets.

PKT_WR_FIFO_SEL

Host Write FIFO Select. In video mode, software shall not program this to Video.

RAW_DATA

Host raw data mode. In this mode, all data is sent exactly as written without any attempt to decode packet headers.

Structs

Register

Constants

CRC_RESET

CS_ENABLE

DSI_HIGH_SPEED_TRANS

DSI_PHY_CLK_DIV

DSI_ULTRA_LOW_POWER

ECC_ENABLE

FIFO_STAT_RESET

HOST_TX_TRIG_SRC

IMM_BTA

PERIPH_RESET

PKT_BTA

PKT_WR_FIFO_SEL

RAW_DATA