Module libtegra::flow::FLOW_CTLR_CPU_CSR_0

Bitfields of the FLOW_CTLR_CPU_<x>_CSR_0 register.

Modules

ENABLE

Power-Gate Enable.

ENABLE_EXT

Specifies what to power off, if ENABLE is set.

EVENT_ENABLE

Generates an event when the Flow Controller exits the halted state.

EVENT_FLAG

Whether Event is Active.

F2C_MPCORE_RST

Whether a Reset was requested for MPCore.

F2P_PWRUP

flow2pmc_pwrup signal.

F2P_REQ

flow2pmc_req valid.

HALT

Whether the CPU is halted.

IMMEDIATE_WAKE

If set, CPU is powered up immediately.

INTR_FLAG

Whether Interrupt is Active.

P2F_ACK

pmc2flow_ack signal.

PWR_STATE

Current state of the Power-Gate State Machine.

SWITCH_CLUSTER

Switches the active cluster when all indicated CPUs reach STANDBY_WFI.

WAIT_EVENT

Whether the CPU is waiting.

WAIT_WFE_BITMAP

All cores indicated in bitmap must be in STANDBY_WFE before CpU power-gating.

WAIT_WFI_BITMAP

All cores indicated in bitmap must be in STANDBY_WFI before CPU power-gating.

Structs

Register

Constants

ENABLE

ENABLE_EXT

EVENT_ENABLE

EVENT_FLAG

F2C_MPCORE_RST

F2P_PWRUP

F2P_REQ

HALT

IMMEDIATE_WAKE

INTR_FLAG

P2F_ACK

PWR_STATE

SWITCH_CLUSTER

WAIT_EVENT

WAIT_WFE_BITMAP

WAIT_WFI_BITMAP