Module libtegra::i2c::I2C_I2C_SL_CNFG_0

Bitfields of the I2C_I2C_SL_CNFG_0 register.

Modules

ACK_LAST_BYTE

Whether the last byte should be acknowledged.

ACK_LAST_BYTE_VALID

Whether the last byte should be acknowledged valid.

ACK_WITHHOLD_EN

Whether the ACK Withhold feature should be enabled.

BUFFER_SIZE

The payload size in bytes.

ENABLE_SL

By writing zero to this field, the slave can be turned off.

FIFO_XFER_EN

Whether data should be communicated through the FIFOs.

NACK

Whether Slave ACK should be disabled.

NEWSL

Whether a new slave should be used.

PKT_MODE_EN

Whether Packet Mode should be enabled.

RESP

Whether slave responses should be sent to the general call address (zero address).

Structs

Register

Constants

ACK_LAST_BYTE

ACK_LAST_BYTE_VALID

ACK_WITHHOLD_EN

BUFFER_SIZE

ENABLE_SL

FIFO_XFER_EN

NACK

NEWSL

PKT_MODE_EN

RESP