Struct libtegra::i2c::I2c

pub struct I2c { /* private fields */ }

Representation of an I2C controller device.

Tegra X1 boards expose 6 different I2C controllers with identical functionality. These are exposed and can be accessed through the constants held by this structure.

Before using any of those, make sure that the bus is properly initialized and ready for data transfers by calling I2c::init with the correct Pin Multiplexing for the controller being done manually in advance.

Implementations of the blocking I2C traits from the embedded-hal crate are provided, if the feature is enabled during build.

Implementations

impl I2c

pub const C1: Self = I2c { clock: &Clock::I2C_1, registers: I2C_1_REGISTERS, }

Representation of the I2C1 controller.

pub const C2: Self = I2c { clock: &Clock::I2C_2, registers: I2C_2_REGISTERS, }

Representation of the I2C2 controller.

pub const C3: Self = I2c { clock: &Clock::I2C_3, registers: I2C_3_REGISTERS, }

Representation of the I2C3 controller.

pub const C4: Self = I2c { clock: &Clock::I2C_4, registers: I2C_4_REGISTERS, }

Representation of the I2C4 controller.

pub const C5: Self = I2c { clock: &Clock::I2C_5, registers: I2C_5_REGISTERS, }

Representation of the I2C5 controller.

pub const C6: Self = I2c { clock: &Clock::I2C_6, registers: I2C_6_REGISTERS, }

Representation of the I2C6 controller.

impl I2c

pub fn send_normal(&self, device: u32, buffer: &[u8]) -> Result<(), Error>

A lower-level API to write a buffer of data to a given I2C device.

Consider using I2c::write if you have no special reason to use this method. If you have, make sure that the first byte of the buffer argument is the desired target register of the device and that the contents of the supplied buffer do not exceed 8 bytes in size.

This method drives the I2C controller to operate in Normal Mode, using 7-bit addressing transactions. Devices are implementation-specific, no guarantees are made that data will be written to actual, valid devices.

pub fn receive_normal(
&self,
device: u32,
buffer: &mut [u8]
) -> Result<(), Error>

A lower-level API to read into a buffer of data from a given I2C device.

Consider using I2c::read if you have no special reason to use this method. If you have, make sure to send a buffer consisting of only the register byte to the desired device with I2c::send_normal first. After that, this method can be used to read up to 8 bytes of data into a supplied buffer.

This method drives the I2C controller to operate in Normal Mode, using 7-bit addressing transactions. Devices are implementation-specific, no guarantees are made that data will be read from actual, valid devices.

pub fn init(&self)

Initializes the I2C device for data transfers.

NOTE: This method must be called once before the I2C device is usable. Further, it is required to do the respective pin configuration before calling this method.

pub fn write(&self, device: u32, register: u8, data: &[u8]) -> Result<(), Error>

Writes a buffer of data to a device register over I²C.

This method drives the I2C controller to operate in Normal Mode, using 7-bit addressing transactions. Devices and their registers are implementation-specific, no guarantees are made that data will be written to actual, valid devices.

pub fn read(
&self,
device: u32,
register: u8,
buffer: &mut [u8]
) -> Result<(), Error>

Reads the contents of a device register over I²C.

This method drives the I2C controller to operate in Normal Mode, using 7-bit addressing transactions. Devices and their registers are implementation-specific, no guarantees are made that data will be read from actual, valid devices.

pub fn write_byte(
&self,
device: u32,
register: u8,
byte: u8
) -> Result<(), Error>

Writes a single data byte to a device register over I²C.

pub fn read_byte(&self, device: u32, register: u8) -> Result<u8, Error>

Reads a single data byte from a device register over I²C.