Module libtegra::kfuse

Driver for the Tegra X1 for the HDCP KFUSE Controller.

See Chapter 27.5 in the Tegra X1 Technical Reference Manual for details.

Description

The KFUSE stores upstream and downstream HDCP keys which are used by the HDMI module to enforce Digital Rights Management (DRM) with hardware protection of the cryptographic secrets.

Initialization

In certain cases, the KFUSE is busy initializing data in the background and thus cannot be used at a point where the host controller is being queried, which may lead to unexpected errors (for example with the TSEC). The wait_until_ready function serves the purpose of preventing that from happening.

Key Copying

HDCP keys are stored encrypted in the KFUSE block. Before starting HDCP, software must copy the keys (576 bytes) from the KFUSE to the HDMI registers.

HDCP

HDCP is usually implemented by running a secure firmware on the Tsec which can interface with KFUSE hardware to do the cryptographic operations on digital content.

Modules

KFUSE_ERRCOUNT_0

Bitfields of the KFUSE_ERRCOUNT_0 register.

KFUSE_KEYADDR_0

Bitfields of the KFUSE_KEYADDR_0 register.

KFUSE_KEYS_0

Bitfields of the KFUSE_KEYS_0 register.

KFUSE_PD_0

Bitfields of the KFUSE_PD_0 register.

KFUSE_STATE_0

Bitfields of the KFUSE_STATE_0 register.

Structs

Registers

Representation of the HDCP KFUSE registers.

Constants

KFUSE_KEY_BUFFER_SIZE

The size of the buffer in words required for reading all encrypted HDMI keys from the KFUSE.

REGISTERS

A pointer to the KFUSE register block that can be accessed by dereferencing it.

Functions

read

Reads the encrypted HDCP keys from the KFUSE into a buffer.

wait_until_ready

Waits until KFUSE is ready to be used.