Module libtegra::pinmux

Driver for Tegra X1 Multi-Purpose Pins and Pin Multiplexing.

See Chapter 9 in the Tegra X1 Technical Reference Manual for details.

Description

Tegra X1 devices can be configured with different I/O functions on particular pins to enable their operation in a variety of different configurations.

Many of the pins on Tegra X1 devices are connected to Multi-Purpose I/O (MPIO) pads. An MPIO can operate in two modes: either acting as a signal for a particular I/O controller, referred to as Special-Function I/O (SFIO) or as a software-controlled General-Purpose I/O function, referred to as Gpio.

Though each MPIO has up to 5 functions (a GPIO function and up to 4 SFIO functions), a given MPIO can only act as a single function at a given point in time. The Pinmux controller in Tegra X1 devices includes the logic and registers to select a particular function for each MPIO.

Configuration

Various get_ and set_ methods are provided to configure PinGrPs (Pin Group Pads) or to query their state. Many devices depend on proper Pin Multiplexing, so this module provides the required functionality to drive the desired pins.

Safety

Many of the configuration methods on a PinGrP are actually considered unsafe because wrong usage of them can cause permanent damage to the hardware which is at the user’s risk.

Modules

PAD_OPTIONS

Pad Control Options for Pinmux registers.

Structs

Registers

Representation of the Pinmux registers.

Enums

PinDrive

State control for the Drive Type control bit as a part of pin configuration.

PinFunction

Pin Functions that can be loaded onto a pin to control its behavior.

PinGrP

Pin Groups on the Tegra X1 SoC that can be customized and variably configured.

PinIo

State control for the I/O direction bit as a part of Pin configuration.

PinIoHv

State control for the I/O High Voltage control bit as a part of pin configuration.

PinLock

State control for the Lock bit as a part of Pin configuration.

PinLpdr

State control for the Base Driver control bit as a part of pin configuration.

PinOd

State control for the OD control bit as a part of pin configuration.

PinPark

State control for the Parking bit as a part of Pin configuration.

PinPull

State control for the Pull bit as a part of Pin configuration.

PinSchmt

State control for the Schmitt Trigger control bit as a part of pin configuration.

PinTristate

State control for the Tristate bit as a part of Pin configuration.

Constants

PINMUX_BASE

Base address for Pinmux registers.

REGISTERS

A pointer to the Pinmux register block that can be accessed by dereferencing it.