Module libtegra::pinmux

Driver for Tegra X1 Multi-Purpose Pins and Pin Multiplexing.

See Chapter 9 in the Tegra X1 Technical Reference Manual for details.

Description

Tegra X1 devices can be configured with different I/O functions on particular pins to allow use in a variety of different configurations.

Many of the pins on Tegra X1 devices are connected to Multi-Purpose I/O (MPIO) pads. An MPIO can operate in two modes: either acting as a signal for a particular I/O controller, referred to as Special-Function I/O (SFIO) or as a software-controlled General-Purpose I/O function, referred to as Gpio. Each MPIO has up to four SFIO PinFunctions as well as being a Gpio.

Though each MPIO has up to 5 functions (a Gpio function and up to 4 SFIO functions), a given MPIO can only act as a single function at a given point in time. The Pinmux controller in Tegra X1 devices includes the logic and registers to select a particular function for each MPIO.

Configuration

Every MPIO on the SoC belongs to a particular PinGrP which provides the functionality for configuration of PinFunction, PinPull state, PinTristate mode, PinIo settings, PinLock controls, PinOd state, and PinEIoHv mode. All of these options can either be configured individually via a respective method or universally via the PinGrP::config method.

use libtegra::pinmux;
use libtegra::pinmux::PinGrP;

// I2C5 Pinmux configuration for the Nintendo Switch.
PinGrP::PwrI2CSclPy3.set_io(pinmux::PinIo::Input);
PinGrP::PwrI2CSdaPy4.set_io(pinmux::PinIo::Input);

Modules

PAD_OPTIONS
Pad Control Options for Pinmux registers.

Structs

Registers
Representation of the Pinmux registers.

Enums

PinEIoHv
Enumeration over possible operation states.

PinFunction
Enumeration over possible Pin Functions.

PinGrP
Enumeration over possible Pin Groups.

PinIo
Enumeration over possible I/O configurations of a pin.

PinLock
Lock control for writing to a pin.

PinOd
Reserved.

PinPull
Enumeration over possible Pin Pull-up/down states.

PinTristate
Tri-State (high-Z) option.

Constants

PINMUX_BASE
Base address for Pinmux registers.

REGISTERS
A pointer to the Pinmux register block that can be accessed by dereferencing it.