Struct libtegra::se::SecurityEngine

pub struct SecurityEngine { /* private fields */ }

Representation of the Security Engine used for cryptographic operations.

Implementations

impl SecurityEngine

pub const SE1: Self = SecurityEngine { registers: SE1_REGISTERS, rsa_keyslot_cache: [rsa::KeyInfo::new(); constants::rsa::KEY_SLOT_COUNT], }

A pointer to the first Security Engine instance.

pub const SE2: Self = SecurityEngine { registers: SE2_REGISTERS, rsa_keyslot_cache: [rsa::KeyInfo::new(); constants::rsa::KEY_SLOT_COUNT], }

A pointer to the second Security Engine instance.

NOTE: Only available with the mariko feature enabled.

impl SecurityEngine

pub fn lock(&self)

Locks the SE down for use from the Secure World.

Only TrustZone clients can access the SE anymore afterwards.

pub fn unlock(&self)

Unlocks the SE for use from the Non-Secure World.

All clients can access the SE afterwards.

pub fn lock_per_key(&self)

Locks down the Security Engine per-key.

pub fn lock_tzram(&self)

Locks down the TZRAM to access from Secure World only.

pub fn lock_context_save(&self)

Locks down the Security Engine to restrict context save operations to Secure World clients.

NOTE: This is available for SE1 and SE2, but only on T210B01 either way.

pub fn disable(&self)

Disables the Security Engine.

pub fn fill_aes_keyslot(&self, slot: u32, key: &[u8])

Fills a given keyslot with the supplied AES key.

This must be done prior to any encryptions using this slot.

pub fn get_aes_key(&self, slot: u32, key: &mut [u8])

Copies a previously loaded AES key out of a given keyslot.

pub fn clear_aes_keyslot(&self, slot: u32)

Clears the data out of a given AES keyslot.

pub fn clear_aes_key_iv(&self, slot: u32)

Clears the IV out of a given AES keyslot.

This affects both, the original and the updated IV values.

pub fn set_encrypted_aes_key(
&self,
dest_slot: u32,
kek_slot: u32,
key: &[u8],
mode: AesMode
) -> Result<(), OperationError>

Loads an encrypted AES key into the given AES keyslot.

The key in the AES keyslot denoted by kek_slot is used to decrypt the supplied key via AES-ECB before loading it into the desired dest_slot in the key table.

pub fn aes_cmac(
&self,
slot: u32,
source: &[u8],
destination: &mut [u8],
mode: AesMode
) -> Result<(), OperationError>

Calculates an AES-CMAC from source to destination.

pub fn aes_ecb_encrypt(
&self,
slot: u32,
source: &[u8; 16],
destination: &mut [u8; 16],
mode: AesMode
) -> Result<(), OperationError>

Encrypts a block of data from source to destination using AES-ECB.

pub fn aes_ecb_decrypt(
&self,
slot: u32,
source: &[u8; 16],
destination: &mut [u8; 16],
mode: AesMode
) -> Result<(), OperationError>

Decrypts a block of data from source to destination using AES-ECB.

pub fn aes_cbc_encrypt(
&self,
slot: u32,
source: &[u8],
destination: &mut [u8],
iv: &[u8; 16],
mode: AesMode
) -> Result<(), OperationError>

Encrypts data from source to destination using AES-CBC.

pub fn aes_cbc_decrypt(
&self,
slot: u32,
source: &[u8],
destination: &mut [u8],
iv: &[u8; 16],
mode: AesMode
) -> Result<(), OperationError>

Decrypts data from source to destination using AES-CBC.

pub fn aes_ctr_encrypt(
&self,
slot: u32,
source: &[u8],
destination: &mut [u8],
iv: &[u8; 16],
mode: AesMode
) -> Result<(), OperationError>

Encrypts data from source to destination using AES-CTR.

pub fn aes_ctr_decrypt(
&self,
slot: u32,
source: &[u8],
destination: &mut [u8],
iv: &[u8; 16],
mode: AesMode
) -> Result<(), OperationError>

Decrypts data from source to destination using AES-CTR.

pub fn clear_rsa_keyslot(&mut self, slot: u32)

Clears all data out of a given RSA keyslot.

pub fn fill_rsa_keyslot(&mut self, slot: u32, modulus: &[u8], exponent: &[u8])

Fills the RSA keyslot using the supplied modulus and exponent data.

This must be done prior to any RSA operations using the selected slot.

pub fn rsa_modular_exponentiate(
&self,
slot: u32,
source: &[u8],
destination: &mut [u8]
) -> Result<(), OperationError>

Computes the modular exponentiation of source ^ exponent (mod n).

Exponent and modulus should have already been loaded into a keyslot prior to calling this method.

pub fn initialize_rng(&self) -> Result<(), OperationError>

Initializes the RNG (Random Numer Generator).

Calling this function is a prerequisite for all functions that use random input sources to perform cryptographic operations.

pub fn generate_random(&self, output: &mut [u8]) -> Result<(), OperationError>

Uses the RNG to fill the given buffer with random bytes.

pub fn set_random_key(&self, slot: u32) -> Result<(), OperationError>

Fills a given key slot with a random key generated by the RNG.

pub fn generate_srk(&self) -> Result<(), OperationError>

Performs a hardware operation to generate the Storage Root Key (SRK).

NOTE: Different entropy sources will lead to different results.

pub fn calculate_sha1(
&self,
source: &[u8],
output: &mut [u8; 20]
) -> Result<(), OperationError>

Calculates a SHA1 hash over a given buffer of data.

pub fn calculate_sha224(
&self,
source: &[u8],
output: &mut [u8; 28]
) -> Result<(), OperationError>

Calculates a SHA224 hash over a given buffer of data.

pub fn calculate_sha256(
&self,
source: &[u8],
output: &mut [u8; 32]
) -> Result<(), OperationError>

Calculates a SHA256 hash over a given buffer of data.

pub fn calculate_sha384(
&self,
source: &[u8],
output: &mut [u8; 48]
) -> Result<(), OperationError>

Calculates a SHA384 hash over a given buffer of data.

pub fn calculate_sha512(
&self,
source: &[u8],
output: &mut [u8; 64]
) -> Result<(), OperationError>

Calculates a SHA512 hash over a given buffer of data.