Module libtegra::spi::SPI_TIMING_REG2_0

Bitfields of the SPI_TIMING_REG2_0 register.

Modules

CS_ACTIVE_BETWEEN_PACKETS_0

Specifies if CS stays active between two packets on CS#0.

CS_ACTIVE_BETWEEN_PACKETS_1

Specifies if CS stays active between two packets on CS#1.

CS_ACTIVE_BETWEEN_PACKETS_2

Specifies if CS stays active between two packets on CS#2.

CS_ACTIVE_BETWEEN_PACKETS_3

Specifies if CS stays active between two packets on CS#3.

CYCLES_BETWEEN_PACKETS_0

Specifies the number of cycles between two packets for communication on CS#0.

CYCLES_BETWEEN_PACKETS_1

Specifies the number of cycles between two packets for communication on CS#1.

CYCLES_BETWEEN_PACKETS_2

Specifies the number of cycles between two packets for communication on CS#2.

CYCLES_BETWEEN_PACKETS_3

Specifies the number of cycles between two packets for communication on CS#3.

RESERVED0

Reserved for future use.

RESERVED1

Reserved for future use.

RESERVED2

Reserved for future use.

RESERVED3

Reserved for future use.

Structs

Register

Constants

CS_ACTIVE_BETWEEN_PACKETS_0

CS_ACTIVE_BETWEEN_PACKETS_1

CS_ACTIVE_BETWEEN_PACKETS_2

CS_ACTIVE_BETWEEN_PACKETS_3

CYCLES_BETWEEN_PACKETS_0

CYCLES_BETWEEN_PACKETS_1

CYCLES_BETWEEN_PACKETS_2

CYCLES_BETWEEN_PACKETS_3

RESERVED0

RESERVED1

RESERVED2

RESERVED3