Module libtegra::tsec

Driver for interfacing with the Tegra Security Co-Processor (TSEC).

Description

The TSEC is a dedicated unit powered by a NVIDIA Falcon microprocessor with crypto extensions. Its purpose is to leverage certain cryptographic tasks done by firmwares signed by NVIDIA into a secure space that cannot be taken over by the host system.

It has three operating modes:

No Secure Mode (NS): Every piece of microcode that is not cryptographically signed by NVIDIA will be executed in this mode. It prevents you from accessing certain registers and may disable physical memory access from code.

Light Secure Mode (LS): In this mode, the Falcon has more privileges than in No Secure Mode, but fewer than in Heavy Secure Mode. This mode leaks some of the internal state to ease up debugging and can only be enabled from Heavy Secure Mode microcode.

Heavy Secure Mode (HS): This mode can be entered by uploading signed microcode and grants the full range of privileges to the microcode. This state essentially turns the Falcon into a black box that doesn’t expose any of its inner workings to, for example, the host system.

Firmware

envytools have proven to be valuable tools when it comes to working with various ISAs used by NVIDIA, including the Falcon processor.

The following examples will use this reference firmware:

mov $r15 0xB0B0B0B0;
mov $r9 0x1100;
iowr I[$r9] $r15;
exit;

It can be assembled with the following command, assuming the code is stored in a file called faucon.asm:

envyas -m falcon -V fuc5 -F crypt faucon.asm -i -o faucon.bin

Firmware alignment

Firmware blobs that should be booted on the Tsec are supposed to be aligned to the boundary denoted by FIRMWARE_ALIGNMENT. This is implied by the Falcon code segment, which consists of 0x100 byte pages. An approach to getting this correct could be:

use libtegra::tsec::{Firmware, FIRMWARE_ALIGNMENT};

/// The firmware blob.
static FW: Firmware<u8, 13> = Firmware::new([
    0xDF, 0xB0, 0xB0, 0xB0, 0xB0,   // mov $r15 0xB0B0B0B0;
    0x49, 0x00, 0x11,               // mov $r9 0x1100;
    0xF6, 0x9F, 0x00,               // iowr I[$r9] $r15;
    0xF8, 0x02,                     // exit;
]);

assert_eq!(&*FW as *const u8 as usize % FIRMWARE_ALIGNMENT, 0);

Executing code

Using our firmware blob and the static variable FW from above, we can load the code onto the TSEC and try to execute it:

use libtegra::tsec::{Firmware, Tsec};

static FW: Firmware<u8, 13> = Firmware::new([
    0xDF, 0xB0, 0xB0, 0xB0, 0xB0,   // mov $r15 0xB0B0B0B0;
    0x49, 0x0, 0x11,                // mov $r9 0x1100;
    0xF6, 0x9F, 0x0,                // iowr I[$r9] $r15;
    0xF8, 0x2,                      // exit;
]);

// Load our firmware onto the TSEC.
Tsec::A.load_firmware(&*FW).unwrap();

// Boot it up!
let mut mailbox0 = 0;
let mut mailbox1 = 0;
unsafe {
    Tsec::A.boot(0, &mut mailbox0, &mut mailbox1).unwrap();
    assert_eq!(mailbox1, 0xB0B0B0B0);
}

Modules

TSEC_BAR0_CTL

Bitfields of the TSEC_BAR0_CTL register.

TSEC_CG

Bitfields of the TSEC_CG register.

TSEC_FALCON_ADDR

Bitfields of the TSEC_FALCON_ADDR register.

TSEC_FALCON_CG2

Bitfields of the TSEC_FALCON_CG2 register.

TSEC_FALCON_CGCTL

Bitfields of the TSEC_FALCON_CGCTL register.

TSEC_FALCON_CMEMBASE

Bitfields of the TSEC_FALCON_CMEMBASE register.

TSEC_FALCON_CPUCTL

Bitfields of the TSEC_FALCON_CPUCTL register.

TSEC_FALCON_CPUCTL_ALIAS

Bitfields of the TSEC_FALCON_CPUCTL_ALIAS register.

TSEC_FALCON_CTXACK

Bitfields of the TSEC_FALCON_CTXACK register.

TSEC_FALCON_CURCTX

Bitfields of the TSEC_FALCON_CURCTX register.

TSEC_FALCON_DEBUG1

Bitfields of the TSEC_FALCON_DEBUG1 register.

TSEC_FALCON_DMACTL

Bitfields of the TSEC_FALCON_DMACTL register.

TSEC_FALCON_DMAINFO_CTL

Bitfields of the TSEC_FALCON_DMAINFO_CTL register.

TSEC_FALCON_DMAPOLL

Bitfields of the TSEC_FALCON_DMAPOLL_FB register.

TSEC_FALCON_DMATRFCMD

Bitfields of the TSEC_FALCON_DMATRFCMD register.

TSEC_FALCON_DMATRFMOFFS

Bitfields of the TSEC_FALCON_DMATRFMOFFS register.

TSEC_FALCON_DMEMAPERT

Bitfields of the TSEC_FALCON_DMEMAPERT register.

TSEC_FALCON_DMEMC

Bitfields of the TSEC_FALCON_DMEMC register.

TSEC_FALCON_ENGCTL

Bitfields of the TSEC_FALCON_ENGCTL register.

TSEC_FALCON_EXCI

Bitfields of the TSEC_FALCON_EXCI register.

TSEC_FALCON_EXTERRSTAT

Bitfields of the TSEC_FALCON_EXTERRSTAT register.

TSEC_FALCON_FHSTATE

Bitfields of the TSEC_FALCON_FHSTATE register.

TSEC_FALCON_HWCFG

Bitfields of the TSEC_FALCON_HWCFG register.

TSEC_FALCON_HWCFG1

Bitfields of the TSEC_FALCON_HWCFG1 register.

TSEC_FALCON_IBRKPT

Bitfields of the TSEC_FALCON_IBRKPTX registers.

TSEC_FALCON_ICD_CMD

Bitfields of the TSEC_FALCON_ICD_CMD register.

TSEC_FALCON_IDLESTATE

Bitfields of the TSEC_FALCON_IDLESTATE register.

TSEC_FALCON_IMCTL

Bitfields of the TSEC_FALCON_IMCTL register.

TSEC_FALCON_IMCTL_DEBUG

Bitfields of the TSEC_FALCON_IMCTL_DEBUG register.

TSEC_FALCON_IMEMC

Bitfields of the TSEC_FALCON_IMEMC register.

TSEC_FALCON_IMEMT

Bitfields of the TSEC_FALCON_IMEMT register.

TSEC_FALCON_IMFILLCTL

Bitfields of the TSEC_FALCON_IMFILLCTL register.

TSEC_FALCON_IMFILLRNG

Bitfields of the TSEC_FALCON_IMFILLRNGX registers.

TSEC_FALCON_IRQDEST

Bitfields of the TSEC_FALCON_IRQDEST register.

TSEC_FALCON_IRQDEST2

Bitfields of the TSEC_FALCON_IRQDEST2 register.

TSEC_FALCON_IRQS

Bitfields of the TSEC_FALCON_IRQXXXX registers.

TSEC_FALCON_ITFEN

Bitfields of the TSEC_FALCON_ITFEN register.

TSEC_FALCON_MTHDCOUNT

Bitfields of the TSEC_FALCON_MTHDCOUNT register.

TSEC_FALCON_MTHDID

Bitfields of the TSEC_FALCON_MTHDID register.

TSEC_FALCON_MTHDPOP

Bitfields of the TSEC_FALCON_MTHDPOP register.

TSEC_FALCON_MTHDRAMSZ

Bitfields of the TSEC_FALCON_MTHDRAMSZ register.

TSEC_FALCON_NXTCTX

Bitfields of the TSEC_FALCON_NXTCTX register.

TSEC_FALCON_PMM

Bitfields of the TSEC_FALCON_PMM register.

TSEC_FALCON_PRIVSTATE

Bitfields of the TSEC_FALCON_PRIVSTATE register.

TSEC_FALCON_SCTL

Bitfields of the TSEC_FALCON_SCTL register.

TSEC_FALCON_SFTRESET

Bitfields of the TSEC_FALCON_SFTRESET register.

TSEC_FALCON_SOFT_MODE

Bitfields of the TSEC_FALCON_SOFT_MODE register.

TSEC_FALCON_SOFT_PM

Bitfields of the TSEC_FALCON_SOFT_PM register.

TSEC_FALCON_SPROT

Bitfields of the TSEC_FALCON_SPROT_X registers.

TSEC_FALCON_SSTAT

Bitfields of the TSEC_FALCON_SSTAT register.

TSEC_FALCON_STACKCFG

Bitfields of the TSEC_FALCON_STACKCFG register.

TSEC_FALCON_SVEC_SPR

Bitfields of the TSEC_FALCON_SVEC_SPR register.

TSEC_FALCON_TRACEIDX

Bitfields of the TSEC_FALCON_TRACEIDX register.

TSEC_FALCON_TRACEPC

Bitfields of the TSEC_FALCON_TRACEPC register.

TSEC_FALCON_WDTMRCTL

Bitfields of the TSEC_FALCON_WDTMRCTL register.

TSEC_SCP_ACL_ERR

Bitfields of the TSEC_SCP_ACL_ERR register.

TSEC_SCP_CFG

Bitfields of the TSEC_SCP_CFG register.

TSEC_SCP_CMD

Bitfields of the TSEC_SCP_CMD register.

TSEC_SCP_CMD_ERR

Bitfields of the TSEC_SCP_CMD_ERR register.

TSEC_SCP_CTL0

Bitfields of the TSEC_SCP_CTL0 register.

TSEC_SCP_CTL1

Bitfields of the TSEC_SCP_CTL1 register.

TSEC_SCP_CTL_LOCK

Bitfields of the TSEC_SCP_CTL_LOCK register.

TSEC_SCP_CTL_PKEY

Bitfields of the TSEC_SCP_CTL_PKEY register.

TSEC_SCP_CTL_SCP

Bitfields of the TSEC_SCP_CTL_SCP register.

TSEC_SCP_CTL_STAT

Bitfields of the TSEC_SCP_CTL_STAT register.

TSEC_SCP_DBG0

Bitfields of the TSEC_SCP_DBG0 register.

TSEC_SCP_DBG1

Bitfields of the TSEC_SCP_DBG1 register.

TSEC_SCP_DBG2

Bitfields of the TSEC_SCP_DBG2 register.

TSEC_SCP_IRQMASK

Bitfields of the TSEC_SCP_IRQMASK register.

TSEC_SCP_IRQSTAT

Bitfields of the TSEC_SCP_IRQSTAT register.

TSEC_SCP_RNG_STAT0

Bitfields of the TSEC_SCP_RNG_STAT0 register.

TSEC_SCP_SEC_ERR

Bitfields of the TSEC_SCP_SEC_ERR register.

TSEC_SCP_STAT0

Bitfields of the TSEC_SCP_STAT0 register.

TSEC_SCP_STAT1

Bitfields of the TSEC_SCP_STAT1 register.

TSEC_SCP_STAT2

Bitfields of the TSEC_SCP_STAT2 register.

TSEC_TEGRA_CTL

Bitfields of the TSEC_TEGRA_CTL register.

TSEC_TFBIF_CTL

Bitfields of the TSEC_TFBIF_CTL register.

TSEC_TFBIF_DBG_STAT0

Bitfields of the TSEC_TFBIF_DBG_STAT0 register.

TSEC_TFBIF_MCCIF_FIFOCTRL

Bitfields of the TSEC_TFBIF_MCCIF_FIFOCTRL register.

TSEC_TFBIF_MCCIF_FIFOCTRL1

Bitfields of the TSEC_TFBIF_MCCIF_FIFOCTRL1 register.

TSEC_TFBIF_REGIONCFG

Bitfields of the TSEC_TFBIF_REGIONCFG register.

TSEC_TFBIF_SPROT_EMEM

Bitfields of the TSEC_TFBIF_SPROT_EMEM register.

TSEC_TFBIF_THI_TRANSPROP

Bitfields of the TSEC_TFBIF_THI_TRANSPROP register.

TSEC_TFBIF_THROTTLE

Bitfields of the TSEC_TFBIF_THROTTLE register.

TSEC_TFBIF_TRANSCFG

Bitfields of the TSEC_TFBIF_TRANSCFG register.

TSEC_TFBIF_WRR_RDP

Bitfields of the TSEC_TFBIF_WRR_RDP register.

TSEC_THI_CONFIG0

Bitfields of the TSEC_THI_CONFIG0 register.

TSEC_THI_CONTEXT_SWITCH

Bitfields of the TSEC_THI_CONTEXT_SWITCH register.

TSEC_THI_CONT_SYNCPT_EOF

Bitfields of the TSEC_THI_CONT_SYNCPT_EOF register.

TSEC_THI_CONT_SYNCPT_L1

Bitfields of the TSEC_THI_CONT_SYNCPT_L1 register.

TSEC_THI_CTXSW

Bitfields of the TSEC_THI_CTXSW register.

TSEC_THI_CTXSW_INCR_SYNCPT

Bitfields of the TSEC_THI_CTXSW_INCR_SYNCPT register.

TSEC_THI_CTXSW_NEXT

Bitfields of the TSEC_THI_CTXSW_NEXT register.

TSEC_THI_DBG_MISC

Bitfields of the TSEC_THI_DBG_MISC register.

TSEC_THI_INCR_SYNCPT

Bitfields of the TSEC_THI_INCR_SYNCPT register.

TSEC_THI_INCR_SYNCPT_CTRL

Bitfields of the TSEC_THI_INCR_SYNCPT_CTRL register.

TSEC_THI_INCR_SYNCPT_ERR

Bitfields of the TSEC_THI_INCR_SYNCPT_ERR register.

TSEC_THI_INT_MASK

Bitfields of the TSEC_THI_INT_MASK register.

TSEC_THI_INT_STATUS

Bitfields of the TSEC_THI_INT_STATUS register.

TSEC_THI_METHOD0

Bitfields of the TSEC_THI_METHOD0 register.

TSEC_THI_SLCG_OVERRIDE_HIGH_A

Bitfields of the TSEC_THI_SLCG_OVERRIDE_HIGH_A register.

TSEC_THI_STREAMID

Bitfields of the TSEC_THI_STREAMIDX register.

TSEC_THI_THI_SEC

Bitfields of the TSEC_THI_THI_SEC register.

Structs

Firmware

A helper structure to align arrays containing Falcon machine code to the expected 0x100 bytes memory alignment for DMA transfers into the code segment.

Registers

Representation of the TSEC registers.

Tsec

Representation of the Tegra Security Co-Processor.

Enums

FalconError

Falcon processor exceptions that may occur when interacting with it from the host system or may be caused by running code.

FalconExceptionClause

Enumeration of potential Falcon processor exception clauses that may occur during code execution on the TSEC.

Constants

FIRMWARE_ALIGNMENT

The alignment a TSEC firmware blob is expected to have.

FIRMWARE_ALIGN_BITS

The alignment bits for TSEC firmware blobs.

TSEC_A_REGISTERS

A pointer to the TSEC-A register block that can be accessed by dereferencing it.

TSEC_B_REGISTERS

A pointer to the TSEC-B register block that can be accessed by dereferencing it.