Module libtegra::tsec

Driver for interfacing with the Tegra Security Co-Processor (TSEC).

Description

The TSEC is a dedicated unit powered by a NVIDIA Falcon microprocessor with crypto extensions.

Falcon Processor

The NVIDIA Falcon microprocessor is an MCU with very limited code and data space, which is why many features in general are not available.

See this article for a more in-depth description of the Falcon LLVM backend and thus the Falcon environment along with its limitations and features.

Within the TSEC, it has special crypto extensions, which makes it usable for execution of critical code within a secure environment.

Firmware

envytools have proven to be valuable tools when it comes to working with various ISAs used by NVIDIA, including the Falcon processor.

The following examples will use this reference firmware:

mov $r15 0xB0B0B0B0;
mov $r12 0x0;
mov $r9 0x1100;
iowr I[$r9] $r15;
exit;

It can be assembled with the following command, assuming the code is stored in a file called faucon.asm:

envyas -m falcon -V fuc5 -F crypt faucon.asm -i -o faucon.bin

Firmware Alignment

Firmware blobs that should be booted on the Tsec are supposed to be aligned to the boundary denoted by FIRMWARE_ALIGNMENT.

use libtegra::tsec::FIRMWARE_ALIGNMENT;

/// A helper that wraps a TSEC firmware blob to ensure correct alignment.
#[repr(align(256))]
struct Firmware<T: Sized> {
    pub value: T,
}

impl<T: Sized> Firmware<T> {
    pub const fn new(value: T) -> Self {
        Firmware { value }
    }
}

/// The firmware blob.
static FAUCON: Firmware<[u8; 15]> = Firmware::new([
    0xDF, 0xB0, 0xB0, 0xB0, 0xB0,   // mov $r15 0xB0B0B0B0;
    0xC, 0x0,                       // mov $r12 0x0;
    0x49, 0x0, 0x11,                // mov $r9 0x1100;
    0xF6, 0x9F, 0x0,                // iowr I[$r9] $r15;
    0xF8, 0x2,                      // exit;
]);

assert_eq!(FAUCON.value.as_ptr() as usize % FIRMWARE_ALIGNMENT, 0);

Code Execution

Using our firmware blob and the static variable FAUCON from above, we can load the code onto the TSEC and try execute it:

use libtegra::tsec::Tsec;

/// The global instance of the TSEC.
const TSEC: Tsec = Tsec::new();

/// A helper that wraps around a TSEC firmware blob to ensure correct alignment.
#[repr(align(256))]
struct Firmware<T: Sized> {
    pub value: T,
}

impl<T: Sized> Firmware<T> {
    pub const fn new(value: T) -> Self {
        Firmware { value }
    }
}

/// The firmware blob.
static FAUCON: Firmware<[u8; 15]> = Firmware::new([
    0xDF, 0xB0, 0xB0, 0xB0, 0xB0,   // mov $r15 0xB0B0B0B0;
    0xC, 0x0,                       // mov $r12 0x0;
    0x49, 0x0, 0x11,                // mov $r9 0x1100;
    0xF6, 0x9F, 0x0,                // iowr I[$r9] $r15;
    0xF8, 0x2,                      // exit;
]);

// Load our Faucon firmware onto the TSEC.
TSEC.load_firmware(&FAUCON.value).unwrap();

// Boot it up!
unsafe {
    TSEC.boot_firmware(0).unwrap();
}

Structs

Registers
Representation of the TSEC registers.

Tsec
Representation of the Tegra Security Co-Processor.

Enums

FalconError
Enumeration of potential Falcon processor exceptions that may occur when interacting with the TSEC.

FalconExceptionClause
Enumeration of potential Falcon processor exception clauses that may occur during code execution on the TSEC.

Constants

FIRMWARE_ALIGNMENT
The alignment a TSEC firmware blob is expected to have.

FIRMWARE_ALIGN_BITS
The alignment bits for TSEC firmware blobs.

REGISTERS
A pointer to the TSEC register block that can be accessed by dereferencing it.