Module libtegra::uart

Driver for the Tegra X1 Universal Asynchronous Receiver/Transmitter Controller.

See Chapter 36 in the Tegra X1 Technical Reference Manual for details.

Description

There are five UARTs available in total. The UARTs A through D, which are identical, are built into Tegra X1 devices and the fifth UART is located in the Audio Processing Engine.

These UARTs support both, 16450 and 16550 compatible modes, however they are always being initialized in 16550 mode. Using them in 16450 mode requires a custom implementation.

UARTs internally use a TX FIFO, which acts as a buffer for data to send, and an RX FIFO, which acts as a buffer for received data, to read from. Hence, UARTs provide a stream of serial data to communicate with other devices.

Initialization

Uarts need to be initialized with a given baud rate before they can be used. The maximum supported rate is 12.5.

use libtegra::uart::Uart;

Uart::A.init(115_200);

Communication

After a Uart was initialized, there are many possibilities how data can be sent:

use core::fmt::Write;

use libtegra::uart::Uart;

let mut uart = Uart::A; // Less typing...

// You can write a slice of bytes...
uart.write(b"48656c6c6f2c20776f726c6421");
// ...just as well as a single byte...
uart.write_byte(b'0');

// ...or even strings with the `Write` trait.
writeln!(&mut uart, "Hello, world!").ok();

let name = "John";
writeln!(&mut uart, "Hello, {}!", name).ok();

Reading data is also supported:

use libtegra::uart::Uart;

let mut uart = Uart::A; // Less typing...

// Read a single byte.
let byte = uart.read_byte();

// Read 10 bytes into a buffer and print the data.
let mut buffer = [0; 10];
uart.read(&mut buffer);

println!("UART A: {}", String::from_utf8_lossy(&buffer));

Flushing

In some cases, you may want to flush the underlying FIFOs:

use libtegra::uart::Uart;

// Clears the data from both, TX FIFO and RX FIFO.
Uart::A.flush();

Modules

UART_ASR_0
Bitfields of the UART_ASR_0 register.

UART_IER_DLAB_0_0
Bitfields of the UART_IER_DLAB_0_0 register.

UART_IIR_FCR_0
Bitfields of the UART_IIR_FCR_0 register.

UART_IRDA_CSR_0
Bitfields of the UART_IRDA_CSR_0 register.

UART_LCR_0
Bitfields of the UART_LCR_0 register.

UART_LSR_0
Bitfields of the UART_LSR_0 register.

UART_MCR_0
Bitfields of the UART_MCR_0 register.

UART_MIE_0
Bitfields of the UART_MIE_0 register.

UART_MSR_0
Bitfields of the UART_MSR_0 register.

UART_RX_FIFO_CFG_0
Bitfields of the UART_RX_FIFO_CFG_0 register.

UART_SPR_0
Bitfields of the UART_SPR_0 register.

UART_THR_DLAB_0_0
Bitfields of the UART_THR_DLAB_0_0 register.

UART_VENDOR_STATUS_0_0
Bitfields of the UART_VENDOR_STATUS_0_0 register.

Structs

Registers
Representation of the UART registers.

Uart
Representation of a UART.

Constants

UART_A_REGISTERS
A pointer to the UART A register block that can be accessed by dereferencing it.

UART_B_REGISTERS
A pointer to the UART B register block that can be accessed by dereferencing it.

UART_C_REGISTERS
A pointer to the UART C register block that can be accessed by dereferencing it.

UART_D_REGISTERS
A pointer to the UART D register block that can be accessed by dereferencing it.

UART_E_REGISTERS
A pointer to the UART (AP)E register block that can be accessed by dereferencing it.