Module libtegra::uart::UART_MCR_0

Bitfields of the UART_MCR_0 register.

Modules

CTS_EN

Whether CTS Hardware Flow Control should be enabled.

DEL_QUAL_CTS_EN

Whether the old qualified CTS in TX state machine should be used.

DTR

Whether DTR should be forced to high or not.

LOOPBK

Whether internal loop back of Serial Out to In should be enabled.

OLD_SIR_DECODE

Whether the old SIR decode path should be used instead of the new one.

OUT1

nOUT1 (Not Used).

OUT2

nOUT2 (Not Used).

RI_POLARITY

Polarity selection bit for RI pin toggling to generate model status interrupt.

RTS

Whether RTS should be forced to high if RTS hardware flow control wasn’t enabled.

RTS_EN

Whether RTS Hardware Flow Control should be enabled.

Structs

Register

Constants

CTS_EN

DEL_QUAL_CTS_EN

DTR

LOOPBK

OLD_SIR_DECODE

OUT1

OUT2

RI_POLARITY

RTS

RTS_EN