Struct libtegra::uart::Uart

pub struct Uart { /* fields omitted */ }

Representation of a UART.

NOTE: Instances of this struct should never be created manually. Refer to the public constants the struct holds, which represent the UARTs A through E.

Methods

impl Uart

pub const A: Self

Representation of UART A.

pub const B: Self

Representation of UART B.

pub const C: Self

Representation of UART C.

pub const D: Self

Representation of UART D.

pub const E: Self

Representation of UART E.

impl Uart

pub fn init(&self, baud_rate: u32)

Initializes the Uart with a given baud rate.

NOTE: This method needs to be called once before a Uart can actually send and receive data. Further, it is required to do the respective pinmux configuration before calling this method.

pub fn read_byte(&self) -> u8

Reads a byte over UART and returns it.

This method blocks until data is available to read.

pub fn read(&self, data: &mut [u8])

Fills a mutable slice of data with bytes read over UART.

This method blocks until the buffer is filled.

pub fn write_byte(&self, byte: u8)

Writes a byte over UART.

This method blocks until data can be transferred.

pub fn write(&self, data: &[u8])

Writes a slice of bytes over UART.

This method blocks until everything was transferred.

pub fn flush(&self)

Flushes the underlying FIFOs of the Uart.

This wipes out the data to read and the data that should be written, so be careful when you use it. In most cases, this method won't be needed.